% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_R6_utils.R
\name{active_binding}
\alias{active_binding}
\title{Helper function to generate active bindings that are read_only}
\usage{
active_binding(value, expr, name)
}
\arguments{
\item{value}{The value attempted to be set}

\item{expr}{The expression to execute when called}

\item{name}{The name of the active binding}
}
\description{
Helper function to generate active bindings that are read_only
}
