% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demography_nordic}
\alias{demography_nordic}
\title{Demography (Nordics)}
\source{
U.S. Census Bureau. (n.d.). \emph{International Database: World Population Estimates and Projections}.
U.S. Department of Commerce. Retrieved
June 15, 2026,
from https://data.census.gov/
}
\description{
This data set population data for 176 geographical regions in
1-year age-groups to allow transformation of the contact matrices into other age cuts.
}
\details{
This product uses the Census Bureau Data API but is not endorsed or certified by the Census Bureau.
}
\author{
Rasmus Skytte Randløv \email{rske@ssi.dk}
}
\keyword{data}
