% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyModelGLM.R
\name{Diseasy-GLM-models}
\alias{Diseasy-GLM-models}
\alias{DiseasyModelGLM}
\alias{DiseasyModelG0}
\alias{DiseasyModelG1}
\title{Simple generalized linear regression models}
\value{
A new instance of the \code{DiseasyModelGLM}, \code{DiseasyModelG0} or
\code{DiseasyModelG1} \link[R6:R6Class]{R6} class.
}
\description{
The \code{DiseasyModelGLM} module implements common structure and functionality to
GLM regression class of models beyond the model structure provided by \code{DiseasyModelRegression}.

Most notably, the model module implements the \verb{$fit_regression()} and \verb{$get_prediction()} methods using
GLM.

\code{diseasy} includes two simple models that uses the \code{DiseasyModelGLM} module:
\code{DiseasyModelG0} and \code{DiseasyModelG1}
These models implements a constant predictor and a exponential model based on the previous 7 and 21 days
of observations, respectively.

When making a custom GLM model, the subclass should implement the \verb{$update_formula()} method.
The \verb{$update_formula()} method should update the formula based on the stratifications.
If the model should flexibly adapt to different stratifications, this method should be implemented.
See \code{DiseasyModelG0} and \code{DiseasyModelG1} for
examples of how this can be done.
}
\examples{
 model <- DiseasyModelG0$new()

 rm(model)
}
\keyword{model-template}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{DiseasyModel}} -> \code{\link[diseasy:DiseasyModelRegression]{DiseasyModelRegression}} -> \code{DiseasyModelGLM}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DiseasyModelGLM-clone}{\code{DiseasyModelGLM$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>DiseasyBaseModule$load_module()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>DiseasyBaseModule$set_moduleowner()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_data'><code>DiseasyModel$get_data()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModelRegression" data-id="get_results"><a href='../../diseasy/html/DiseasyModelRegression.html#method-DiseasyModelRegression-get_results'><code>DiseasyModelRegression$get_results()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModelRegression" data-id="initialize"><a href='../../diseasy/html/DiseasyModelRegression.html#method-DiseasyModelRegression-initialize'><code>DiseasyModelRegression$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModelRegression" data-id="plot"><a href='../../diseasy/html/DiseasyModelRegression.html#method-DiseasyModelRegression-plot'><code>DiseasyModelRegression$plot()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelGLM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelGLM-clone}{}}}
\subsection{\code{DiseasyModelGLM$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyModelGLM$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{DiseasyModel}} -> \code{\link[diseasy:DiseasyModelRegression]{DiseasyModelRegression}} -> \code{\link[diseasy:DiseasyModelGLM]{DiseasyModelGLM}} -> \code{DiseasyModelB_} -> \code{DiseasyModelG0}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DiseasyModelG0-initialize}{\code{DiseasyModelG0$new()}}
    \item \href{#method-DiseasyModelG0-clone}{\code{DiseasyModelG0$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>DiseasyBaseModule$load_module()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>DiseasyBaseModule$set_moduleowner()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_data'><code>DiseasyModel$get_data()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModelRegression" data-id="get_results"><a href='../../diseasy/html/DiseasyModelRegression.html#method-DiseasyModelRegression-get_results'><code>DiseasyModelRegression$get_results()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModelRegression" data-id="plot"><a href='../../diseasy/html/DiseasyModelRegression.html#method-DiseasyModelRegression-plot'><code>DiseasyModelRegression$plot()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelG0-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelG0-initialize}{}}}
\subsection{\code{DiseasyModelG0$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyModelG0$new(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{parameters sent to \code{DiseasyModelG_} \link[R6:R6Class]{R6} constructor}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new instance of the \code{DiseasyModelG1} \link[R6:R6Class]{R6} class.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelG0-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelG0-clone}{}}}
\subsection{\code{DiseasyModelG0$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyModelG0$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{DiseasyModel}} -> \code{\link[diseasy:DiseasyModelRegression]{DiseasyModelRegression}} -> \code{\link[diseasy:DiseasyModelGLM]{DiseasyModelGLM}} -> \code{DiseasyModelB_} -> \code{DiseasyModelG1}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DiseasyModelG1-initialize}{\code{DiseasyModelG1$new()}}
    \item \href{#method-DiseasyModelG1-clone}{\code{DiseasyModelG1$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>DiseasyBaseModule$load_module()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>DiseasyBaseModule$set_moduleowner()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_data'><code>DiseasyModel$get_data()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModelRegression" data-id="get_results"><a href='../../diseasy/html/DiseasyModelRegression.html#method-DiseasyModelRegression-get_results'><code>DiseasyModelRegression$get_results()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModelRegression" data-id="plot"><a href='../../diseasy/html/DiseasyModelRegression.html#method-DiseasyModelRegression-plot'><code>DiseasyModelRegression$plot()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelG1-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelG1-initialize}{}}}
\subsection{\code{DiseasyModelG1$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyModelG1$new(...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{...}}{parameters sent to \code{DiseasyModelG_} \link[R6:R6Class]{R6} constructor}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A new instance of the \code{DiseasyModelG1} \link[R6:R6Class]{R6} class.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelG1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelG1-clone}{}}}
\subsection{\code{DiseasyModelG1$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyModelG1$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
