% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasy-package.R
\docType{package}
\name{diseasy-package}
\alias{diseasy}
\alias{diseasy-package}
\title{diseasy: Ensemble Disease Modelling Made Easy}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Modules, models and tools to create ensembles of disease models through modular construction and combinatorics. Using the data pipeline provided by the 'diseasystore' package, 'diseasy' enables ensembles of disease models to be run across multiple diseases. By using the global activity matrices from the 'contactdata' package, the models can run on almost all countries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ssi-dk/diseasy}
  \item Report bugs at \url{https://github.com/ssi-dk/diseasy/issues}
}

}
\author{
\strong{Maintainer}: Rasmus Skytte Randløv \email{rske@ssi.dk} (\href{https://orcid.org/0000-0002-5860-3838}{ORCID})

Authors:
\itemize{
  \item Lasse Engbo Christiansen \email{lsec@ssi.dk} (\href{https://orcid.org/0000-0001-5019-1931}{ORCID})
}

Other contributors:
\itemize{
  \item Carsten Kirkeby \email{ckir@sund.ku.dk} (\href{https://orcid.org/0000-0001-9292-5526}{ORCID}) [contributor]
  \item Frederik Plesner Lyngse \email{frederik.lyngse@sund.ku.dk} (\href{https://orcid.org/0000-0002-0584-6115}{ORCID}) [contributor]
  \item Adam Mielke \email{admi@dtu.dk} (\href{https://orcid.org/0000-0001-8605-2167}{ORCID}) [contributor]
  \item Marcus Munch Grünewald \email{mmgr@ssi.dk} (\href{https://orcid.org/0009-0006-8090-406X}{ORCID}) [contributor]
  \item Kaare Græsbøll \email{kagr@ssi.dk} (\href{https://orcid.org/0000-0002-6258-8212}{ORCID}) [contributor]
  \item Statens Serum Institut, SSI [copyright holder, funder]
}

}
\keyword{internal}
