% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_linters.R
\name{diseasy_linters}
\alias{diseasy_linters}
\alias{diseasy_code_linters}
\alias{non_ascii_linter}
\alias{todo_linter}
\title{The custom linters of \code{diseasy}}
\usage{
diseasy_code_linters()

non_ascii_linter()

todo_linter()
}
\value{
A list of linters
}
\description{
The custom linters of \code{diseasy}
}
\examples{
  diseasy_code_linters()
}
