% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyDBModule.R
\name{DiseasyDBModule}
\alias{DiseasyDBModule}
\title{DiseasyBaseModule with DB connection}
\description{
TODO
}
\seealso{
\link{DiseasyBaseModule}, \link[SCDB:get_connection]{SCDB::get_connection}

\link[SCDB:get_table]{SCDB::get_table}
}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasyDBModule}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{slice_ts}}{(\code{Date})\cr
The timestamp to slice database on. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyDBModule-new}{\code{DiseasyDBModule$new()}}
\item \href{#method-DiseasyDBModule-set_slice_ts}{\code{DiseasyDBModule$set_slice_ts()}}
\item \href{#method-DiseasyDBModule-clone}{\code{DiseasyDBModule$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyDBModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyDBModule-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \code{DiseasyDBModule} \link[R6:R6Class]{R6} class.
This module is typically not constructed directly but rather through derived classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyDBModule$new(conn = NULL, slice_ts = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conn}}{(\verb{DBI::dbConnect object})\cr
A database connection object (inherits from DBIConnection)}

\item{\code{slice_ts}}{(\code{Date} or \code{character})\cr
Date to slice the database on. See \code{\link[SCDB:get_table]{SCDB::get_table()}}}

\item{\code{...}}{parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the \code{DiseasyDBModule} \link[R6:R6Class]{R6} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyDBModule-set_slice_ts"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyDBModule-set_slice_ts}{}}}
\subsection{Method \code{set_slice_ts()}}{
Set the slice_ts to get data for
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyDBModule$set_slice_ts(slice_ts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slice_ts}}{(\code{Date} or \code{character})\cr
Date to slice the database on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyDBModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyDBModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyDBModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
