% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyEnsemble.R
\name{DiseasyEnsemble-generics}
\alias{DiseasyEnsemble-generics}
\alias{print.DiseasyEnsemble}
\alias{summary.DiseasyEnsemble}
\alias{plot.DiseasyEnsemble}
\title{Standard generics for \code{DiseasyEnsemble} objects}
\usage{
\method{print}{DiseasyEnsemble}(x, width = 200, ...)

\method{summary}{DiseasyEnsemble}(object, ...)

\method{plot}{DiseasyEnsemble}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{width}{(\code{integer(1)})\cr
The maximum number of characters to print.}

\item{...}{further arguments passed to or from other methods.}

\item{object}{an object for which a summary is desired.}
}
\value{
\code{NULL} (called for side effects)
}
\description{
Standard generics for \code{DiseasyEnsemble} objects
}
\examples{
\dontshow{if (rlang::is_installed("duckdb")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  observables <- DiseasyObservables$new(
    diseasystore = DiseasystoreSeirExample,
    conn = DBI::dbConnect(duckdb::duckdb())
  )

  # Create a DiseasyEnsemble object
  ensemble <- combineasy(
    model_templates = list(DiseasyModelG0, DiseasyModelG1)
    modules = observables
  )

  print(ensemble)

  summary(ensemble)

  plot(ensemble)

  rm(ensemble, observables)
\dontshow{\}) # examplesIf}
}
