% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyModel.R
\name{DiseasyModel}
\alias{DiseasyModel}
\title{Meta module for the models}
\description{
TODO
}
\seealso{
\link{DiseasyObservables}

\link{DiseasyActivity}

\link{DiseasyObservables}

\link{DiseasySeason}
}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasyModel}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{activity}}{(\code{diseasy::activity})\cr
The local copy of an activity module. Read-only.}

\item{\code{observables}}{(\code{diseasy::DiseasyObservables})\cr
The local copy of an DiseasyObservables module. Read-only.}

\item{\code{season}}{(\code{diseasy::season})\cr
The local copy of an season module. Read-only.}

\item{\code{parameters}}{(\code{list()})\cr
The parameters used in the model. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyModel-new}{\code{DiseasyModel$new()}}
\item \href{#method-DiseasyModel-load_module}{\code{DiseasyModel$load_module()}}
\item \href{#method-DiseasyModel-get_results}{\code{DiseasyModel$get_results()}}
\item \href{#method-DiseasyModel-get_training_data}{\code{DiseasyModel$get_training_data()}}
\item \href{#method-DiseasyModel-clone}{\code{DiseasyModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModel-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyModel} \link[R6:R6Class]{R6} class.
This module is typically not constructed directly but rather through \verb{DiseasyModel*} classes, such as \link{DiseasyModelG0}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModel$new(
  activity = FALSE,
  observables = FALSE,
  season = FALSE,
  label = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activity, observables, season}}{(\code{boolean} or \verb{R6::R6Class instance})\cr
If a boolean is given, it dictates whether to load a new instance module of this class\cr
If an instance of the module is provided instead, this instance is cloned to the new \code{DiseasyModel} instance\cr
Default is FALSE.}

\item{\code{label}}{(\code{character})\cr
A human readable label for the model instance}

\item{\code{...}}{parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The \code{DiseasyModel} is the main template that the individual models should inherit from since this defines the
set of methods the later framework expects from each model. In addition, it provides the main interface with
the other modules of the framework
}

\subsection{Returns}{
A new instance of the \code{DiseasyModel} \link[R6:R6Class]{R6} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModel-load_module"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModel-load_module}{}}}
\subsection{Method \code{load_module()}}{
Loads a copy of the provided module into the module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModel$load_module(mod, reference_mod)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod}}{(\code{boolean} or \verb{R6::R6Class instance})\cr
If a boolean is given, it dictates whether to load a new instance module of the reference_mod class\cr
If an instance of the same class as reference_mod is provided instead, this instance is cloned to the field
with the same name class(reference_mod).}

\item{\code{reference_mod}}{(\code{R6::R6Class})\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The methods allows the setting of the internal module instances after the \code{DiseasyModel} instance is created.
}

\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModel-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModel-get_results}{}}}
\subsection{Method \code{get_results()}}{
The primary method used to request model results of a given observable at a given aggregation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModel$get_results(
  observable,
  prediction_length,
  quantiles = NULL,
  aggregation = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{observable}}{(\code{character})\cr
The observable to provide prediction for. Must match observable in \code{DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{\code{observable}}{(\code{character})\cr
The observable to provide prediction for. Must match observable in \code{DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{\code{prediction_length}}{(\code{numeric})\cr
The number of days to predict.
The prediction start is defined by \code{last_queryable_date} of the \code{DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{\code{quantiles}}{(\code{list}(\code{numeric}))\cr
Default NULL.
If given, results are returned at the quantiles given}

\item{\code{aggregation}}{(\code{list}(\code{quosures}))\cr
Default NULL.
If given, expressions in aggregation evaluated to give the aggregation level.}

\item{\code{aggregation}}{(\code{list}(\code{quosures}))\cr
Default NULL.
If given, expressions in aggregation evaluated to give the aggregation level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble with predictions at the level specified by aggregation level. In addition to aggregation columns, the output has the columns:\cr   date (\code{Date}) specifying the date of the prediction\cr   realization_id (\code{character}) giving a unique id for each realization in the ensemble\cr   model (\code{character}) the name (classname) of the model used to provide the prediction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModel-get_training_data"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModel-get_training_data}{}}}
\subsection{Method \code{get_training_data()}}{
A method that returns training data for the models based on the model value of \code{training_length} and
the \code{last_queryable_date} of the \code{DiseasyObservables} module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModel$get_training_data(observable, aggregation = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{observable}}{(\code{character})\cr
The observable to provide prediction for. Must match observable in \code{DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{\code{observable}}{(\code{character})\cr
The observable to provide prediction for. Must match observable in \code{DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{\code{aggregation}}{(\code{list}(\code{quosures}))\cr
Default NULL.
If given, expressions in aggregation evaluated to give the aggregation level.}

\item{\code{aggregation}}{(\code{list}(\code{quosures}))\cr
Default NULL.
If given, expressions in aggregation evaluated to give the aggregation level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The output of \code{DiseasyObservables$get_observation} constrained to the training period.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
