% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyModelB_.R
\name{DiseasyModelB0}
\alias{DiseasyModelB0}
\title{Model module for the b0 reference model}
\description{
TODO
}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{diseasy::DiseasyModel}} -> \code{\link[diseasy:DiseasyRegression]{diseasy::DiseasyRegression}} -> \code{\link[diseasy:DiseasyModelBRM]{diseasy::DiseasyModelBRM}} -> \code{diseasy::NA} -> \code{DiseasyModelB0}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyModelB0-new}{\code{DiseasyModelB0$new()}}
\item \href{#method-DiseasyModelB0-clone}{\code{DiseasyModelB0$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_training_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_training_data'><code>diseasy::DiseasyModel$get_training_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="load_module"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-load_module'><code>diseasy::DiseasyModel$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyRegression" data-id="get_results"><a href='../../diseasy/html/DiseasyRegression.html#method-DiseasyRegression-get_results'><code>diseasy::DiseasyRegression$get_results()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelB0-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelB0-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyModelB_} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelB0$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters sent to \code{DiseasyModelB_} \link[R6:R6Class]{R6} constructor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelB0-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelB0-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelB0$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
