% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyModelBRM.R
\name{DiseasyModelBRM}
\alias{DiseasyModelBRM}
\title{Meta module for the BRM class of models}
\description{
TODO
}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{diseasy::DiseasyModel}} -> \code{\link[diseasy:DiseasyRegression]{diseasy::DiseasyRegression}} -> \code{DiseasyModelBRM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyModelBRM-clone}{\code{DiseasyModelBRM$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_training_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_training_data'><code>diseasy::DiseasyModel$get_training_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="load_module"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-load_module'><code>diseasy::DiseasyModel$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyRegression" data-id="get_results"><a href='../../diseasy/html/DiseasyRegression.html#method-DiseasyRegression-get_results'><code>diseasy::DiseasyRegression$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyRegression" data-id="initialize"><a href='../../diseasy/html/DiseasyRegression.html#method-DiseasyRegression-initialize'><code>diseasy::DiseasyRegression$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelBRM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelBRM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelBRM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
