% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyModelG_.R
\name{DiseasyModelG1}
\alias{DiseasyModelG1}
\title{Model module for the g1 reference model}
\description{
TODO
}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{diseasy::DiseasyModel}} -> \code{\link[diseasy:DiseasyRegression]{diseasy::DiseasyRegression}} -> \code{\link[diseasy:DiseasyModelGLM]{diseasy::DiseasyModelGLM}} -> \code{diseasy::NA} -> \code{DiseasyModelG1}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyModelG1-new}{\code{DiseasyModelG1$new()}}
\item \href{#method-DiseasyModelG1-clone}{\code{DiseasyModelG1$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_training_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_training_data'><code>diseasy::DiseasyModel$get_training_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="load_module"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-load_module'><code>diseasy::DiseasyModel$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyRegression" data-id="get_results"><a href='../../diseasy/html/DiseasyRegression.html#method-DiseasyRegression-get_results'><code>diseasy::DiseasyRegression$get_results()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelG1-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelG1-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyModelG1} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelG1$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters sent to \code{DiseasyModelG_} \link[R6:R6Class]{R6} constructor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelG1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelG1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelG1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
