% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyModelRegression.R
\name{DiseasyModelRegression}
\alias{DiseasyModelRegression}
\title{Base module for the regression class of models}
\value{
A new instance of the \code{DiseasyModelRegression} \link[R6:R6Class]{R6} class.
}
\description{
The \code{DiseasyModelRegression} module implements common structure and functionality to regression class of models
beyond the model structure provided by \code{?DiseasyModel}.

Most notably, the model module implements the \verb{$get_results()} method.
This implementation requires the subclass to implement the \verb{$fit_regression()}, \verb{$get_prediction()} and
\verb{$update_formula()} methods.

The \verb{$fit_regression()} method should fit the regression model to the training data.
In the case of a GLM model, this would be a call to \code{stats::glm}.

The \verb{$get_prediction()} method should predict the future values of the observable.
In the case of a GLM model, this would be a call to \code{stats::predict}.

The \verb{$update_formula()} method should update the formula based on the stratifications.
If the model should flexibly adapt to different stratifications, this method should be implemented.
See \code{?DiseasyModelGLM} and \code{?DiseasyModelBRM} for examples of how this can be done.
}
\examples{
  # This module cannot be constructed directly but should instead be used to
  # inherit from when creating a new model class.
}
\seealso{
\link[stats:family]{stats::family}, \link[stats:formula]{stats::as.formula}, \link{DiseasyModelG0}, \link{DiseasyModelG1}, \link{DiseasyModelB0}, \link{DiseasyModelG1}

\link{DiseasyObservables}
}
\keyword{model-template-builder}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{diseasy::DiseasyModel}} -> \code{DiseasyModelRegression}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{formula}}{(\code{formula})\cr
The base formula of the module. Stratification features extend this base formula. Read-only.}

\item{\code{family}}{(\code{family})\cr
The family used in the regression fit (see \code{glm} or \code{brms}). Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyModelRegression-new}{\code{DiseasyModelRegression$new()}}
\item \href{#method-DiseasyModelRegression-get_results}{\code{DiseasyModelRegression$get_results()}}
\item \href{#method-DiseasyModelRegression-plot}{\code{DiseasyModelRegression$plot()}}
\item \href{#method-DiseasyModelRegression-clone}{\code{DiseasyModelRegression$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_data'><code>diseasy::DiseasyModel$get_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelRegression-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelRegression-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyModelRegression} \link[R6:R6Class]{R6} class.
This module is typically not constructed directly but rather through \verb{DiseasyModel*} classes,
such as \link{DiseasyModelG0}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelRegression$new(formula, family, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{(\code{character})\cr
A \code{character} string that is passed to \code{stats::as.formula} via \code{glue} (see details).}

\item{\code{family}}{(\code{family})\cr
\code{stats::family} object passed to the regression call.}

\item{\code{...}}{parameters sent to \code{DiseasyModel} \link[R6:R6Class]{R6} constructor.}

\item{\code{parameters}}{(\verb{named list()})\cr
\itemize{
\item \code{training_length} (\verb{named numeric(3)})\cr   The number of days that should be included in the training splits of the data for the model.   Allowed splits are: "training", "testing", and "validation".
}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The observable will change at run time and we therefore cannot define a static formula.
We can use "\{observable\}" in our formula which will then be translated at run time.
For example, if the requested observable is "n_hospital" and the formula is "\{observable\} ~ 1",
then at run time, the formula will translate to "n_hospital ~ 1".

Furthermore the stratification can also change at run time, so the model should incorporate a
\code{update_formula(stratification)} function that accounts for changes in stratification.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelRegression-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelRegression-get_results}{}}}
\subsection{Method \code{get_results()}}{
The primary method used to request model results of a given observable at a given stratification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelRegression$get_results(
  observable,
  prediction_length,
  quantiles = NULL,
  stratification = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{observable}}{(\code{character})\cr The observable to provide data or prediction for.}

\item{\code{prediction_length}}{(\code{numeric})\cr The number of days to predict. The prediction start is defined by \code{last_queryable_date} of the \code{?DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{\code{quantiles}}{(\code{list}(\code{numeric}))\cr If given, results are returned at the quantiles given.}

\item{\code{stratification}}{(\code{list}(\code{quosures}) or \code{NULL})\cr Use \code{rlang::quos(...)} to specify stratification. If given, expressions in stratification evaluated to give the stratification level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble} \link[tibble:tibble]{tibble::tibble} with predictions at the level specified by stratification level. In addition to stratification columns, the output has the columns:\cr   date (\code{Date}) specifying the date of the prediction\cr   realization_id (\code{character}) giving a unique id for each realization in the ensemble\cr   model (\code{character}) the name (classname) of the model used to provide the prediction.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelRegression-plot"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelRegression-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the predictions from the current model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelRegression$plot(
  observable,
  prediction_length,
  stratification = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{observable}}{(\code{character})\cr The observable to provide data or prediction for.}

\item{\code{prediction_length}}{(\code{numeric})\cr The number of days to predict. The prediction start is defined by \code{last_queryable_date} of the \code{?DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{\code{stratification}}{(\code{list}(\code{quosures}) or \code{NULL})\cr Use \code{rlang::quos(...)} to specify stratification. If given, expressions in stratification evaluated to give the stratification level.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelRegression-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelRegression-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
