% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyObservables.R
\name{DiseasyObservables}
\alias{DiseasyObservables}
\title{Observable handler}
\value{
A new instance of the \code{DiseasyBaseModule} \link[R6:R6Class]{R6} class.
}
\description{
TODO
}
\examples{
  # Create observables module using the Google COVID-19 data
  obs <- DiseasyObservables$new(case_definition = "Google COVID-19",
                                conn = DBI::dbConnect(RSQLite::SQLite()))

  # See available observables
  print(obs$available_observables)
  print(obs$available_aggregations)

  # Get data for one observable
  obs$get_observation("n_hospital",
                      start_date = as.Date("2020-03-01"),
                      end_date = as.Date("2020-04-01"))

  rm(obs)
}
\seealso{
\link[SCDB:get_table]{SCDB::get_table}

\link[SCDB:get_table]{SCDB::get_table}
}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasyObservables}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{case_definition}}{(\code{character})\cr
The set case_definition to get DiseasyObservables for. Read-only.}

\item{\code{start_date}}{(\code{Date})\cr
The start date of the study period. Read-only.}

\item{\code{end_date}}{(\code{Date})\cr
The end date of the study period. Read-only.}

\item{\code{last_queryable_date}}{(\code{Date})\cr
The latest date that can be queried. Read-only.}

\item{\code{ds}}{(\verb{Diseasystore*})\cr
The currently loaded diseasystore which provides the features. Read-only.}

\item{\code{available_observables}}{(\code{character})\cr
The currently available observables in the loaded diseasystore. Read-only.}

\item{\code{available_aggregations}}{(\code{character})\cr
The currently available aggregations in the loaded diseasystore. Read-only.}

\item{\code{slice_ts}}{(\code{Date})\cr
The timestamp to slice database on. Read-only.}

\item{\code{conn}}{(\code{DBIConnection})\cr
The connection to the database on. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyObservables-new}{\code{DiseasyObservables$new()}}
\item \href{#method-DiseasyObservables-set_case_definition}{\code{DiseasyObservables$set_case_definition()}}
\item \href{#method-DiseasyObservables-set_last_queryable_date}{\code{DiseasyObservables$set_last_queryable_date()}}
\item \href{#method-DiseasyObservables-set_study_period}{\code{DiseasyObservables$set_study_period()}}
\item \href{#method-DiseasyObservables-set_slice_ts}{\code{DiseasyObservables$set_slice_ts()}}
\item \href{#method-DiseasyObservables-get_observation}{\code{DiseasyObservables$get_observation()}}
\item \href{#method-DiseasyObservables-describe}{\code{DiseasyObservables$describe()}}
\item \href{#method-DiseasyObservables-finalize}{\code{DiseasyObservables$finalize()}}
\item \href{#method-DiseasyObservables-clone}{\code{DiseasyObservables$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyObservables-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyObservables-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyObservables} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyObservables$new(
  case_definition = NULL,
  start_date = NULL,
  end_date = NULL,
  last_queryable_date = NULL,
  conn = NULL,
  slice_ts = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{case_definition}}{(\code{character})\cr
A character string that controls which feature store to get data from.}

\item{\code{start_date}}{(\code{Date})\cr
Study period start (default values for get_observation).}

\item{\code{end_date}}{(\code{Date})\cr
Study period end (default values for get_observation).}

\item{\code{last_queryable_date}}{(\code{Date})\cr
Enforce a limit on data that can be pulled (not after this date).}

\item{\code{conn}}{(\code{DBIConnection})\cr
A database connection object (inherits from DBIConnection)}

\item{\code{slice_ts}}{(\code{Date} or \code{character})\cr
Date to slice the database on. See \code{\link[SCDB:get_table]{SCDB::get_table()}}}

\item{\code{...}}{parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyObservables-set_case_definition"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyObservables-set_case_definition}{}}}
\subsection{Method \code{set_case_definition()}}{
Set the case definition to get DiseasyObservables for.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyObservables$set_case_definition(case_definition)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{case_definition}}{(\code{character})\cr
Text label of the disease to get DiseasyObservables for.\cr
Must match case definition implemented in \code{featurestore} package.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyObservables-set_last_queryable_date"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyObservables-set_last_queryable_date}{}}}
\subsection{Method \code{set_last_queryable_date()}}{
Enforce a limit on data that can be pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyObservables$set_last_queryable_date(last_queryable_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{last_queryable_date}}{(\code{Date})\cr
DiseasyObservables module will not return data after this date.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyObservables-set_study_period"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyObservables-set_study_period}{}}}
\subsection{Method \code{set_study_period()}}{
Set the (default) time period to get observations from.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyObservables$set_study_period(start_date, end_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{(\code{Date})\cr
Start date to get DiseasyObservables for (including).}

\item{\code{end_date}}{(\code{Date})\cr
End date to get DiseasyObservables for (including).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyObservables-set_slice_ts"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyObservables-set_slice_ts}{}}}
\subsection{Method \code{set_slice_ts()}}{
Set the slice_ts to get data for
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyObservables$set_slice_ts(slice_ts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slice_ts}}{(\code{Date} or \code{character})\cr
Date to slice the database on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyObservables-get_observation"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyObservables-get_observation}{}}}
\subsection{Method \code{get_observation()}}{
Retrieve an "observable" in the data set corresponding to the set case_definition.\cr
By default, the internal values for start_date and end_date are used to return data,
but these can be overwritten.\cr
The results are cached for faster retrieval at subsequent calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyObservables$get_observation(
  observable,
  aggregation = NULL,
  start_date = self \%.\% start_date,
  end_date = self \%.\% end_date
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{observable}}{(\code{character})\cr
The requested observable. Should follow the pattern 'n_*'.}

\item{\code{aggregation}}{(\code{list}(\code{quosures}))\cr
Default \code{NULL}. If given, expressions in aggregation evaluated to give the aggregation level.\cr
Use rlang::quos(...) to specify aggregation.}

\item{\code{start_date}}{(\code{Date})\cr
Start date to get DiseasyObservables for (including).}

\item{\code{end_date}}{(\code{Date})\cr
End date to get DiseasyObservables for (including).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If the observable is found, the function returns the corresponding data at the aggregation level.\cr
Otherwise, the function fails and lists the available DiseasyObservables for the case_definition.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyObservables-describe"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyObservables-describe}{}}}
\subsection{Method \code{describe()}}{
Prints a human readable report of the internal state of the module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyObservables$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyObservables-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyObservables-finalize}{}}}
\subsection{Method \code{finalize()}}{
Handles the clean-up of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyObservables$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyObservables-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyObservables-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyObservables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
