% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyRegion.R
\name{DiseasyRegion}
\alias{DiseasyRegion}
\title{Diseasy' region (spatial) handler}
\value{
A new instance of the \code{DiseasyRegion} \link[R6:R6Class]{R6} class.
}
\description{
TODO

See the vignette("diseasy-region") for examples of use.
}
\examples{
  # TODO
}
\keyword{functional-module}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasyRegion}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyRegion-new}{\code{DiseasyRegion$new()}}
\item \href{#method-DiseasyRegion-describe}{\code{DiseasyRegion$describe()}}
\item \href{#method-DiseasyRegion-clone}{\code{DiseasyRegion$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyRegion-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyRegion-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyRegion} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyRegion$new(level = c("national", "regional"), filter = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{(\code{character(1)})\cr
The geographic level of interest:}

\item{\code{filter}}{(\code{data.frame})\cr
The subset to consider for the model.}

\item{\code{...}}{parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyRegion-describe"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyRegion-describe}{}}}
\subsection{Method \code{describe()}}{
Prints a human readable report of the internal state of the module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyRegion$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyRegion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyRegion-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyRegion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
