% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasySeason.R
\name{DiseasySeason}
\alias{DiseasySeason}
\title{Diseasy' season handler}
\value{
A new instance of the \code{DiseasySeason} \link[R6:R6Class]{R6} class.
}
\description{
The \code{DiseasySeason} module is responsible for implementing various models for the seasonal dependency of the
diseases.
The module implements a number season models with different functional forms.
Models for season are either extracted from the module through \verb{get_*} functions or the module is configured to
use these models internally through \verb{use_*} functions whereafter the model can be accessed through \verb{$model_t()}
and \verb{$model_date()}.
Each season model has varying number of parameters. See documentation for each for details.

See the vignette("diseasy-season") for examples of use.
}
\examples{
  # Season module with an constant season
  s1 <- DiseasySeason$new()

  x <- 0:365
  plot(x, purrr::map_dbl(x, s1$model_t))

  # Season module with an consine season
  s2 <- DiseasySeason$new(reference_date = Sys.Date())
  s2$use_cosine_season()
  plot(x, purrr::map_dbl(x, s2$model_t))

  rm(s1, s2)
}
\seealso{
\link{DiseasyObservables}
}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasySeason}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{reference_date}}{(\code{Date})\cr
The reference date of the season models. Read-only.}

\item{\code{model_t}}{(\code{function})\cr
The model currently being used in the module (days past reference date). Read-only.}

\item{\code{model_date}}{(\code{function})\cr
The model currently being used in the module (date of interest). Read-only.}

\item{\code{available_season_models}}{(\code{character})\cr
The list of available season models}

\item{\code{observables}}{(\code{diseasy::DiseasyObservables})\cr
The local copy of an DiseasyObservables module. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasySeason-new}{\code{DiseasySeason$new()}}
\item \href{#method-DiseasySeason-set_reference_date}{\code{DiseasySeason$set_reference_date()}}
\item \href{#method-DiseasySeason-set_scale}{\code{DiseasySeason$set_scale()}}
\item \href{#method-DiseasySeason-get_season_model}{\code{DiseasySeason$get_season_model()}}
\item \href{#method-DiseasySeason-use_season_model}{\code{DiseasySeason$use_season_model()}}
\item \href{#method-DiseasySeason-get_constant_season}{\code{DiseasySeason$get_constant_season()}}
\item \href{#method-DiseasySeason-use_constant_season}{\code{DiseasySeason$use_constant_season()}}
\item \href{#method-DiseasySeason-get_cosine_season}{\code{DiseasySeason$get_cosine_season()}}
\item \href{#method-DiseasySeason-use_cosine_season}{\code{DiseasySeason$use_cosine_season()}}
\item \href{#method-DiseasySeason-get_covid_season_v1}{\code{DiseasySeason$get_covid_season_v1()}}
\item \href{#method-DiseasySeason-use_covid_season_v1}{\code{DiseasySeason$use_covid_season_v1()}}
\item \href{#method-DiseasySeason-get_covid_season_v2}{\code{DiseasySeason$get_covid_season_v2()}}
\item \href{#method-DiseasySeason-use_covid_season_v2}{\code{DiseasySeason$use_covid_season_v2()}}
\item \href{#method-DiseasySeason-describe}{\code{DiseasySeason$describe()}}
\item \href{#method-DiseasySeason-clone}{\code{DiseasySeason$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="finalize"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-finalize'><code>diseasy::DiseasyBaseModule$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasySeason} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$new(reference_date = NULL, observables = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reference_date}}{(\code{Date})\cr
Date the season modifier is computed relatively to.}

\item{\code{observables}}{(\verb{R6::R6Class instance})\cr
A instance of \code{DiseasyObservables} are needed for some season models.}

\item{\code{...}}{parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-set_reference_date"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-set_reference_date}{}}}
\subsection{Method \code{set_reference_date()}}{
Sets the reference_date for the \code{DiseasySeason} module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$set_reference_date(reference_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reference_date}}{(\code{Date})\cr
Date the season modifier is computed relatively to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-set_scale"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-set_scale}{}}}
\subsection{Method \code{set_scale()}}{
Sets the scale for the active season model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$set_scale(scale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{(\code{numeric})\cr
The scale of the season effect (relative to climate normal).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-get_season_model"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-get_season_model}{}}}
\subsection{Method \code{get_season_model()}}{
Retrieves the specified season model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$get_season_model(model_name, dots = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the season_model to use (calls the equivalent $get_<model_name>()).}

\item{\code{dots}}{(\code{list})\cr
Named list of arguments that will be passed at dot-ellipsis to the season model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-use_season_model"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-use_season_model}{}}}
\subsection{Method \code{use_season_model()}}{
Sets the \code{DiseasySeason} module to use the specified season model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$use_season_model(model_name, dots = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{Name of the season_model to use (calls the equivalent $use_<model_name>()).}

\item{\code{dots}}{(\code{list})\cr
Named list of arguments that will be passed at dot-ellipsis to the season model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-get_constant_season"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-get_constant_season}{}}}
\subsection{Method \code{get_constant_season()}}{
Retrieves the season model with a constant value (1).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$get_constant_season()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-use_constant_season"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-use_constant_season}{}}}
\subsection{Method \code{use_constant_season()}}{
Sets the season module to use a constant value (1).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$use_constant_season()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-get_cosine_season"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-get_cosine_season}{}}}
\subsection{Method \code{get_cosine_season()}}{
Retrieves the season model with a cosine relationship.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$get_cosine_season(peak = 20.09946, scale = 0.5726693)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peak}}{(\code{numeric})\cr
Sets the period of maximal activity (days past new-year).
By default, risk of infection is antiphase with the DMI climate normal of the maximum daily temperature.}

\item{\code{scale}}{(\code{numeric})\cr Sets the scale of the season model. The scale is the percent wise difference between most active and least active period.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-use_cosine_season"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-use_cosine_season}{}}}
\subsection{Method \code{use_cosine_season()}}{
Sets the \code{DiseasySeason} module to use a cosine model for season.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$use_cosine_season(peak = 20.09946, scale = 0.5726693)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{peak}}{(\code{numeric})\cr
Sets the period of maximal activity (days past new-year).
By default, risk of infection is antiphase with the DMI climate normal of the maximum daily temperature.}

\item{\code{scale}}{(\code{numeric})\cr Sets the scale of the season model. The scale is the percent wise difference between most active and least active period.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-get_covid_season_v1"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-get_covid_season_v1}{}}}
\subsection{Method \code{get_covid_season_v1()}}{
Retrieves the first version of the COVID-19 season model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$get_covid_season_v1(scale = 0.4825524)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{(\code{numeric})\cr Sets the scale of the season model. The scale is the percent wise difference between most active and least active period.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-use_covid_season_v1"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-use_covid_season_v1}{}}}
\subsection{Method \code{use_covid_season_v1()}}{
Sets the \code{DiseasySeason} module to use the first version of the covid 19 season model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$use_covid_season_v1(scale = 0.4825524)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{(\code{numeric})\cr Sets the scale of the season model. The scale is the percent wise difference between most active and least active period.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-get_covid_season_v2"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-get_covid_season_v2}{}}}
\subsection{Method \code{get_covid_season_v2()}}{
Retrieves the second version of the COVID-19 season model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$get_covid_season_v2(scale = 0.5042782)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{(\code{numeric})\cr Sets the scale of the season model. The scale is the percent wise difference between most active and least active period.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-use_covid_season_v2"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-use_covid_season_v2}{}}}
\subsection{Method \code{use_covid_season_v2()}}{
Sets the \code{DiseasySeason} module to use the second version of the COVID-19 season model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$use_covid_season_v2(scale = 0.5042782)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{(\code{numeric})\cr Sets the scale of the season model. The scale is the percent wise difference between most active and least active period.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-describe"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-describe}{}}}
\subsection{Method \code{describe()}}{
Prints a human readable report of the internal state of the module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasySeason-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasySeason-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasySeason$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
