% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dk_activity_units}
\alias{dk_activity_units}
\title{Danish activity "units"}
\description{
This data set contains the "units" of activity that was developed for the Danish COVID-19 modelling efforts.
Each "activity unit" specifies the amount of activity for the four arenas: Home, Work, School, Other.
By combining these units of activity, a complete, granular picture of the societal activity is formed
(see dk_reference_scenario for details).
}
\details{
These units are developed to cover the proposed and realized scenarios of societal restrictions during the COVID-19
pandemic in Denmark. If other scenarios are to be tested, new "activity units" will (likely) have to be developed.
We supply the script that generates these units with the package (see data-raw/dk-activity-units.R for details).
}
\author{
Lasse Engbo Christiansen \email{lsec@ssi.dk}

Carsten Kirkeby \email{ckir@sund.ku.dk}

Frederik Plesner Lyngse \email{frederik.lyngse@sund.ku.dk}

Adam Mielke \email{admi@dtu.dk}

Rasmus Skytte Randløv \email{rske@ssi.dk}
}
\keyword{data}
