% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dk_reference_scenario}
\alias{dk_reference_scenario}
\title{Danish reference scenario (SARS-CoV-2 pandemic)}
\description{
This data set contains the combination of "activity units" (see dk_activity_units for details) that form a
description of the societal activity in Denmark through out the COVID-19 pandemic.
}
\details{
The data set consists of a \code{tibble} with four columns:
\itemize{
\item \code{date}
\item \code{opening}
\item \code{closing}
\item \code{social_distance_work}
}

\code{date} indicates the date of the activity change
\code{opening} indicates the "activity unit" that is added to the scenario on the given date
\code{closing} indicates the "activity unit" that is removed from the scenario on the given date
\code{social_distance_work} indicates the reduction in transmission risk in the "Work" arena due to social distancing
}
\author{
Lasse Engbo Christiansen \email{lsec@ssi.dk}
}
\keyword{data}
