% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_contact_basis.R
\name{generate_contact_basis}
\alias{generate_contact_basis}
\title{Generate contact-basis data}
\usage{
generate_contact_basis(regions = NULL)
}
\arguments{
\item{regions}{(\code{character()})\cr The geographic regions of interest. If \code{NULL}, all regions are returned.}
}
\value{
A named list of contact-basis objects, using the current package data
structure.
}
\description{
Generate the contact-basis object used by \code{diseasy} from \code{contactdata}
contact matrices and country-level demography.
}
\examples{
\dontrun{
contact_basis <- generate_contact_basis()
}
}
\keyword{data-generators}
