% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_demography_nuts3.R
\name{generate_demography_nuts3}
\alias{generate_demography_nuts3}
\title{Generate NUTS3 demography data}
\usage{
generate_demography_nuts3(regions = NULL, cache = FALSE, output_nuts = FALSE)
}
\arguments{
\item{regions}{(\code{character()})\cr The geographic regions of interest. If \code{NULL}, all regions are returned.}

\item{cache}{(\code{logical(1)})\cr
Passed to \code{\link[eurostat:get_eurostat]{eurostat::get_eurostat()}}.}

\item{output_nuts}{(\code{logical(1)})\cr
Should we also return the list of NUTS3 codes?}
}
\value{
A named list with elements \code{demography_nuts3} and \code{nuts}.
}
\description{
Generate NUTS3 demography data and the corresponding NUTS region lookup
from Eurostat dataset \code{demo_r_pjangrp3}.
}
\examples{
\dontrun{
nuts_data <- generate_demography_nuts3(regions = c("DK", "FI", "IS", "NO", "SE"))
}
}
\keyword{data-generators}
