% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_hypoexponential.R
\name{HypoDist}
\alias{HypoDist}
\alias{dhypo}
\alias{phypo}
\alias{qhypo}
\alias{rhypo}
\title{The hypoexponential distribution}
\usage{
dhypo(x, shape = 1, rate = rep(1, shape))

phypo(q, shape = 1, rate = rep(1, shape), lower.tail = TRUE)

qhypo(p, shape = 1, rate = rep(1, shape), lower.tail = TRUE)

rhypo(x, shape = 1, rate = rep(1, shape))
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{shape}{(\code{numeric(1)})\cr
shape parameter}

\item{rate}{an alternative way to specify the scale.}

\item{lower.tail}{logical; if TRUE (default), probabilities are
    \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}
}
\value{
\itemize{
\item \code{dhypo} gives the density.
\item \code{phypo} gives the distribution function.
\item \code{qhypo} gives the quantile function.
}
}
\description{
Density, distribution function and quantile function for the hypoexponential distribution with
parameters shape and rate.
}
\details{
Note that \code{qhypo} has poorer precision (\eqn{~ 10^{-3}}) than \code{dhypo} and \code{phypo} (\eqn{~ 10^{-6}}).
}
\examples{
  dhypo(1:10, shape = 2)
  dhypo(1:10, shape = 2, rate = c(1, 2))

  phypo(0.75, shape = 2)
  phypo(0.75, shape = 2, rate = c(1, 2))
  phypo(0.75, shape = 2, rate = c(1, 2), lower.tail = FALSE)

  qhypo(0.75, shape = 2)
  qhypo(0.75, shape = 2, rate = c(1, 2))
  qhypo(0.75, shape = 2, rate = c(1, 2), lower.tail = FALSE)
}
