% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasy-package.R
\docType{data}
\name{mg_activity_units}
\alias{mg_activity_units}
\title{mg_activity_units}
\description{
mg_activity_units
}
\references{
\url{...}
}
\author{
My name \email{blahblah@roxygen.org}
}
\keyword{data}
