% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasy-package.R
\docType{data}
\name{mg_reference_scenario}
\alias{mg_reference_scenario}
\title{mg_reference_scenario}
\description{
mg_reference_scenario
}
\references{
\url{...}
}
\author{
My name \email{blahblah@roxygen.org}
}
\keyword{data}
