% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyBaseModule.R
\name{plot}
\alias{plot}
\alias{plot.DiseasyBaseModule}
\title{Plotting method for \code{diseasy} modules}
\usage{
\method{plot}{DiseasyBaseModule}(obj, ...)
}
\arguments{
\item{obj}{(\verb{Diseasy*})\cr
The module to generate visualisation for.}

\item{...}{Parameters sent to the specific plotting methods.
See the \verb{$plot()} method for each module}
}
\value{
NULL (called for side effects)
}
\description{
Visualise \code{diseasy} modules.
}
\examples{
  immunity <- DiseasyImmunity$new()
  immunity$plot()

  immunity$set_exponential_waning(target = "hospitalisation")
  immunity$plot()

  rm(immunity)
}
