% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_plotter.R
\name{prediction_plotter}
\alias{prediction_plotter}
\title{A unified plotter for \code{model} \link[R6:R6Class]{R6} classes.}
\usage{
prediction_plotter(
  models,
  observable,
  prediction_length,
  aggregation = NULL,
  context_length = 0,
  na.rm = TRUE
)
}
\arguments{
\item{models}{(\code{list}(\code{model}))\cr
A list of initialized \code{model} \link[R6:R6Class]{R6} classes.}

\item{observable}{(\code{character})\cr
The observable to provide prediction for. Must match observable in \code{DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{prediction_length}{(\code{numeric})\cr
The number of days to predict.
The prediction start is defined by \code{last_queryable_date} of the \code{DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{aggregation}{(\code{list}(\code{quosures}))\cr
Default NULL.
If given, expressions in aggregation evaluated to give the aggregation level.}

\item{context_length}{(\code{numeric})\cr
Default \code{NULL}.
The number of days to plot prior to prediction. If \code{NULL}, it uses the largest \code{training_length} of the models.}

\item{na.rm}{(\code{boolean})\cr
Boolean that controls whether stratification with NA should be filtered out.}
}
\value{
A unified \code{ggplot2} figure.
}
\description{
A unified plotter for \code{model} \link[R6:R6Class]{R6} classes.
}
