% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyBaseModule.R
\name{print}
\alias{print}
\alias{print.DiseasyBaseModule}
\title{Printing method for \code{diseasy} modules}
\usage{
\method{print}{DiseasyBaseModule}(x, ...)
}
\arguments{
\item{x}{(\verb{Diseasy*})\cr
The module to get information for.}

\item{...}{Parameters sent to the specific printing methods.
See the \verb{$print()} method for each module}
}
\value{
\code{NULL} (called for side effects)
}
\description{
Summarise \code{diseasy} modules.
}
\examples{
# Examples to come
}
