% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_R6_utils.R
\name{printr}
\alias{printr}
\title{cat printing with default new line ' ' @param ...  The normal input to cat ' @param file Path of an output file to append the output to ' @param sep The separator given to cat ' @export}
\usage{
printr(..., file = "/dev/null", sep = "")
}
\description{
cat printing with default new line ' ' @param ...  The normal input to cat ' @param file Path of an output file to append the output to ' @param sep The separator given to cat ' @export
}
