% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_R6_utils.R
\name{read_only_error}
\alias{read_only_error}
\title{Helper function to produce a "read only" error}
\usage{
read_only_error(field)
}
\arguments{
\item{field}{The name of the field that is read only}
}
\description{
Helper function to produce a "read only" error
}
