% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seir_example_data}
\alias{seir_example_data}
\title{seir_example_data}
\description{
This data set contains a the simulation output from a SEIR model to use when testing and developing models.
}
\details{
The underlying model is a SEIR model (\code{?DiseasyModelOdeSeir}) with 2 exposed compartments, 1 infected compartment
and 1 recovered compartment. The model uses an exposed period of 2.1 days and a infectious period of 4.5 days.
The model uses Denmark for its population and activity scenario with the "baseline" activity unit as the level of
activity. The overall infection risk is set to be 0.025 and initially, 0.2 \% of the population is newly exposed
(i.e. placed in the E1 compartments).

The model outputs included in the dataset is in the form of a tibble with the following columns:
\itemize{
\item \code{date}: The time for the output.
\item \code{age_group}: The age group for the output.
\item \code{n_infected}: The direct output of the SEIR model ( = (I1 * L) / 4.5 days).
\item \code{n_positive_simple}: A realisation of the number of test-positives with a 65\% test probability.
\item \code{n_positive_simple}: A realisation of the number of test-positives with a overall 65\% test probability but with
a reduced test-probability in the weekend.
\item \code{n_admission}: A realisation of the number of admissions to hospital delayed relative to incidence and
accounting for a age dependent risk of admission.
}
}
\author{
Rasmus Skytte Randl\\u00F8v \email{rske@dtu.dk}
}
\keyword{data}
