% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjacency_meta_nordic_nuts3}
\alias{adjacency_meta_nordic_nuts3}
\title{Meta Social Connectedness Index as adjacency for NUTS 3 regions (Nordics)}
\source{
Meta Social Connectedness Index. Data accessed from the Humanitarian Data Exchange.
Retrieved \verb{r format(attr(adjacency_meta_nordic_nuts3, "creation_date"), "\%B \%d, \%Y", tz = "Europe/Copenhagen")}.
See \url{https://data.humdata.org/dataset/social-connectedness-index} and
\url{https://ai.meta.com/ai-for-good/docs/methodology-social-connectedness-index/}.
}
\description{
This data set contains a long-form adjacency object derived from the Meta Social Connectedness Index
for NUTS 3 regions in then Nordic countries. The object is formatted for use with \code{DiseasyRegions} and
\code{DiseasyRegionsNuts}.
}
\details{
The data set consists of a \code{data.frame} with three columns:
\itemize{
\item \code{from}: The origin NUTS 3 region.
\item \code{to}: The destination NUTS 3 region.
\item \code{adjacency}: The contacts from \code{from} allocated to \code{to}.
}

The adjacency values are two-way contact between regions and should be interpreted a type of "infection-flow" matrix.
}
\keyword{data}
