% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineasy.R
\name{combineasy}
\alias{combineasy}
\title{Easily combine modules and parameters to form a model ensemble}
\usage{
combineasy(model_generators, modules = NULL, parameters = NULL)
}
\arguments{
\item{model_generators}{list(\code{DiseasyModel})\cr
Provide a list of generators (factories) that will be initialized with modules and parameters}

\item{modules}{\code{tibble}\cr
The combination of modules to load into the model instances (generated by \code{tidyr}::\code{expand_grid})}

\item{parameters}{(\code{tibble})\cr
The combination of parameters to set in the model instances (generated by \code{tidyr}::\code{expand_grid}).}
}
\description{
Easily combine modules and parameters to form a model ensemble
}
\seealso{
\link[tidyr:expand_grid]{tidyr::expand_grid}
}
