% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_basis_nordic}
\alias{contact_basis_nordic}
\title{Contact basis (Nordics)}
\source{
U.S. Census Bureau. (n.d.). \emph{International Database: World Population Estimates and Projections}.
U.S. Department of Commerce.
Retrieved August 19, 2026.
from https://data.census.gov/

Gruson H (2023). "contactdata: Social Contact Matrices for 152
Countries." R package version 1.0.0,
\url{https://CRAN.R-project.org/package=contactdata}.

Prem K, Cook AR, Jit M (2017). "Projecting social contact
matrices in 152 countries using contact surveys and demographic
data." \emph{PLoS Computational Biology}, \emph{13}(9), e1005697.
doi:10.1371/journal.pcbi.1005697
\url{https://doi.org/10.1371/journal.pcbi.1005697}.

Prem K, van Zandvoort K, Klepac P, Eggo RM, Davies NG, Group
CCW, Cook AR, Jit M (2021). "Projecting contact matrices in 177
geographical regions: An update and comparison with empirical
data for the COVID-19 era." \emph{PLoS Computational Biology},
\emph{17}(7), e1009098. doi:10.1371/journal.pcbi.1009098
\url{https://doi.org/10.1371/journal.pcbi.1009098}.
}
\description{
This data set contains contact matrices and population data for 176 geographical regions.
The contact matrices uses data provided by the \code{contactdata} package while the population
data is provided by the US Census Bureau.
}
\details{
The contact matrices are provided for each of the four arenas: "Work", "Home", "School", "Other" in
5-year age groups.

The \code{diseasy} SEIR models are configured to use a specific transformation of the contact matrices.
To make the definition of "contact" matrix more clear, lets start with the definitions from this paper:
https://www.medrxiv.org/content/10.1101/2020.02.16.20023754v2.full.pdf

m_ij the raw contact matrix elements from age group i to age group j\cr
c_ij the reciprocal contact matrix elements\cr
w_i  the proportion of population that fall into age group i\cr

The \code{diseasy} SEIR models are configured to use a symmetric, weighted set of
contacts matrices where the elements are 0.5 * (c_ij * w_j + c_ji * w_i)

Since the \code{contactdata} package gives the number of contacts directly (in their
framework, denoted as X_ij), we transform to the above elements through these intermediaries: \cr
m_ij = X_ij / w_j\cr
c_ij = m_ij * w_i = X_ij * w_i / w_j\cr

The population data includes the proportion of the population in these 5-year age groups as well as
information of 1-year age-groups to allow transformation of the contact matrices into other age cuts.

This product uses the Census Bureau Data API but is not endorsed or certified by the Census Bureau.
}
\author{
Rasmus Skytte Randløv \email{rske@ssi.dk}

Lasse Engbo Christiansen \email{lsec@ssi.dk}
}
\keyword{data}
