% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseasy-package.R
\docType{package}
\name{diseasy-package}
\alias{diseasy}
\alias{diseasy-package}
\title{diseasy: Disease Ensemble Modelling Made Easy}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package facilitates creation of an ensemble of models through modular construction and combinatorics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ssi-dk/diseasy}
  \item \url{https://ssi-dk.github.io/diseasy/}
  \item Report bugs at \url{https://github.com/ssi-dk/diseasy/issues}
}

}
\author{
\strong{Maintainer}: Rasmus Skytte Randløv \email{rske@ssi.dk} (\href{https://orcid.org/0000-0002-5860-3838}{ORCID})

Authors:
\itemize{
  \item Marcus Munch Grünewald \email{mmgr@ssi.dk} (\href{https://orcid.org/0009-0006-8090-406X}{ORCID})
  \item Lasse Engbo Christiansen \email{lsec@ssi.dk} (\href{https://orcid.org/0000-0001-5019-1931}{ORCID})
  \item Kaare Græsbøll \email{kagr@ssi.dk} (\href{https://orcid.org/0000-0002-6258-8212}{ORCID})
  \item Sofia Myrup Otero \email{smot@ssi.dk}
}

Other contributors:
\itemize{
  \item Statens Serum Institut, SSI [copyright holder, funder]
}

}
\keyword{internal}
