% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_non_ascii_linter.R, R/diseasy-package.R
\name{diseasy_linters}
\alias{diseasy_linters}
\alias{non_ascii_linter}
\alias{diseasy_code_linters}
\title{The custom linters of \code{diseasy}}
\usage{
non_ascii_linter()

diseasy_code_linters()
}
\value{
A list of linters
}
\description{
The custom linters of \code{diseasy}
}
\examples{
  diseasy_code_linters()
}
