% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyActivity.R
\name{DiseasyActivity}
\alias{DiseasyActivity}
\title{Configure activity (contact) scenarios}
\value{
A new instance of the \code{DiseasyActivity} \link[R6:R6Class]{R6} class.
}
\description{
The \code{DiseasyActivity} module is responsible for handling the societal activity.
By providing a \code{contact_basis}, \code{activity_units} and a \code{scenario} to the module, the
module provides activity matrices throughout the scenario (with flexible age-groupings)
as well as the overall "activity" level in society.

The \code{contact_basis} contains information about:
- contact matrices (contact rates between age groups)
- population (size and proportion of population in age groups)

The \code{activity_units} projects restrictions, guidelines and other changes in activity into smaller "units" that are
independently "opened" or "closed". Opening (closing) a activity unit means the activity described in the unit is
(in)active.

The \code{scenario} contains information on when different \code{activity_units} are opened and closed.

If no scenario is provided, the module will provide non-informative activity information:
- Openness is always 1.
- The contact matrices are uniform and, when summed, the largest eigenvalue is 1.

See vignette("diseasy-activity") for examples of use.
}
\examples{
  # Activity module with Danish reference scenario
  act <- DiseasyActivity$new(base_scenario = "dk_reference",
                             activity_units = dk_activity_units,
                             contact_basis = contact_basis_nordic \%.\% DK)

  # Get contact matrices
  contact_matrices <- act$get_scenario_activities()


  # Configuring a custom scenario in another country (using Danish activity units)
  scenario <- data.frame(date = as.Date(character(0)),
                         opening = character(0),
                         closing = character(0)) |>
    dplyr::add_row(date = as.Date("2020-01-01"), opening = "baseline",      closing = NA) |>
    dplyr::add_row(date = as.Date("2020-03-12"), opening = NA,              closing = "baseline") |>
    dplyr::add_row(date = as.Date("2020-03-12"), opening = "lockdown_2020", closing = NA)

  act$set_contact_basis(contact_basis_nordic \%.\% SE) # Use the "Sweden" contact_basis
  act$set_activity_units(dk_activity_units)
  act$change_activity(scenario)

  # Get societal "openness"
  openness <- act$get_scenario_openness()

  rm(act)
}
\seealso{
\link{dk_activity_units}
}
\keyword{functional-module}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{DiseasyBaseModule}} -> \code{DiseasyActivity}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{scenario_matrix}}{(\code{matrix} \code{array})\cr
A reduced view of the internal state of restrictions (showing only used activity units). Read-only.}

    \item{\code{risk_matrix}}{(\code{matrix} \code{array})\cr
A reduced view of the internal state of overall risk multipliers. Read-only.}

    \item{\code{contact_basis}}{(\code{list(list())})\cr A nested list with all the needed information for the contact_basis\cr * \code{counts} contains the age stratified contact counts across the arenas of the basis   (e.g. 'work', 'home', 'school', 'other')\cr * \code{proportion} contains a list of the proportion of population in each age-group\cr * \code{demography} contains a \code{data.frame} with the columns\cr   * \code{age} (\code{integer()}) 1-year age group\cr   * \code{population} (\code{numeric()}) size of population in age group\cr   * \code{proportion} (\code{numeric()}) proportion of total population in age group\cr * \code{description} contains information about the source of the contact basis. Read only.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DiseasyActivity-initialize}{\code{DiseasyActivity$new()}}
    \item \href{#method-DiseasyActivity-set_activity_units}{\code{DiseasyActivity$set_activity_units()}}
    \item \href{#method-DiseasyActivity-set_contact_basis}{\code{DiseasyActivity$set_contact_basis()}}
    \item \href{#method-DiseasyActivity-change_activity}{\code{DiseasyActivity$change_activity()}}
    \item \href{#method-DiseasyActivity-change_risk}{\code{DiseasyActivity$change_risk()}}
    \item \href{#method-DiseasyActivity-crop_scenario}{\code{DiseasyActivity$crop_scenario()}}
    \item \href{#method-DiseasyActivity-get_scenario_activities}{\code{DiseasyActivity$get_scenario_activities()}}
    \item \href{#method-DiseasyActivity-get_scenario_openness}{\code{DiseasyActivity$get_scenario_openness()}}
    \item \href{#method-DiseasyActivity-get_scenario_contacts}{\code{DiseasyActivity$get_scenario_contacts()}}
    \item \href{#method-DiseasyActivity-rescale_contacts_to_rates}{\code{DiseasyActivity$rescale_contacts_to_rates()}}
    \item \href{#method-DiseasyActivity-reset_scenario}{\code{DiseasyActivity$reset_scenario()}}
    \item \href{#method-DiseasyActivity-plot}{\code{DiseasyActivity$plot()}}
    \item \href{#method-DiseasyActivity-describe}{\code{DiseasyActivity$describe()}}
    \item \href{#method-DiseasyActivity-clone}{\code{DiseasyActivity$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>DiseasyBaseModule$load_module()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-initialize}{}}}
\subsection{\code{DiseasyActivity$new()}}{
  Creates a new instance of the \code{DiseasyActivity} \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$new(
  base_scenario = "closed",
  activity_units = NULL,
  contact_basis = NULL,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{base_scenario}}{(\code{character(1)})\cr
Baseline scenario. Must be either fully "open" or "closed" or "dk_reference"}
      \item{\code{activity_units}}{(\code{list(list())})\cr A nested list of all possible 'units' of activity that can be opened or closed.}
      \item{\code{contact_basis}}{(\code{list(list())})\cr A nested list with all the needed information for the contact_basis\cr * \code{counts} contains the age stratified contact counts across the arenas of the basis   (e.g. 'work', 'home', 'school', 'other')\cr * \code{proportion} contains a list of the proportion of population in each age-group\cr * \code{demography} contains a \code{data.frame} with the columns\cr   * \code{age} (\code{integer()}) 1-year age group\cr   * \code{population} (\code{numeric()}) size of population in age group\cr   * \code{proportion} (\code{numeric()}) proportion of total population in age group\cr * \code{description} contains information about the source of the contact basis.}
      \item{\code{...}}{Parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-set_activity_units"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-set_activity_units}{}}}
\subsection{\code{DiseasyActivity$set_activity_units()}}{
  Sets the list of all possible "units" of activity that can be opened or closed
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$set_activity_units(activity_units)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{activity_units}}{(\code{list(list())})\cr A nested list of all possible 'units' of activity that can be opened or closed.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Details}{
    Each element in the activity_units list should be a list with the following elements:
\itemize{
\item activity: a programmatic short hand for activity (character, snake_case),
\item label: a human readable label for activity (character),
\item home:   numeric/vector with number(s) in [0, 1]
\item work:   numeric/vector with number(s) in [0, 1]
\item school: numeric/vector with number(s) in [0, 1]
\item other:  numeric/vector with number(s) in [0, 1]
\item risk:   numeric greater than zero
}

If a single number is provider, the number is applied across all age-groups
If a vector is provided, the vector must match the number of age groups in the contact_basis
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-set_contact_basis"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-set_contact_basis}{}}}
\subsection{\code{DiseasyActivity$set_contact_basis()}}{
  Sets the contact matrix basis the activity is computed from
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$set_contact_basis(contact_basis)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{contact_basis}}{(\code{list(list())})\cr A nested list with all the needed information for the contact_basis\cr * \code{counts} contains the age stratified contact counts across the arenas of the basis   (e.g. 'work', 'home', 'school', 'other')\cr * \code{proportion} contains a list of the proportion of population in each age-group\cr * \code{demography} contains a \code{data.frame} with the columns\cr   * \code{age} (\code{integer()}) 1-year age group\cr   * \code{population} (\code{numeric()}) size of population in age group\cr   * \code{proportion} (\code{numeric()}) proportion of total population in age group\cr * \code{description} contains information about the source of the contact basis.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Details}{
    Loads the given \code{contact_basis} into the module
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-change_activity"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-change_activity}{}}}
\subsection{\code{DiseasyActivity$change_activity()}}{
  Adds the specified openings and closings to the scenario
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$change_activity(date, opening = NA, closing = NA)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{date}}{(\code{Date()} | \code{data.frame})\cr
Either a vector of dates when activity changes or a data.frame with columns 'date', 'opening' and 'closing'}
      \item{\code{opening}}{(\code{character()})\cr
Names of activities to open on given date.
Omitted if \code{data.frame} is given to date argument}
      \item{\code{closing}}{(\code{character()})\cr
Names of activities to close on given date.
Omitted if \code{data.frame} is given to date argument}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-change_risk"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-change_risk}{}}}
\subsection{\code{DiseasyActivity$change_risk()}}{
  Sets the overall risk of types of activity
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$change_risk(date, type = NULL, risk = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{date}}{(\code{Date()})\cr
Dates where risk changes.
The first argument can also be a data.frame with columns "date", "type" and "risk"}
      \item{\code{type}}{(\code{character(1)})\cr
Name of activity type to change. Must be in "work", "school", "home" and "other"}
      \item{\code{risk}}{(\code{numeric(1)})\cr
Relative risk for the given type from the given date}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-crop_scenario"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-crop_scenario}{}}}
\subsection{\code{DiseasyActivity$crop_scenario()}}{
  Helper function to crop the scenario matrix in time
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$crop_scenario(first_date = NULL, last_date = NULL)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{first_date}}{(\code{Date(1)})\cr
New first date in scenario. The column for the lasted prior date will be new first column}
      \item{\code{last_date}}{(\code{Date(1)})\cr
All columns after this date are deleted}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-get_scenario_activities"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-get_scenario_activities}{}}}
\subsection{\code{DiseasyActivity$get_scenario_activities()}}{
  Return \code{list} containing the active activity units on dates where there are changes.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$get_scenario_activities()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    (\code{list})\cr
The activity units active in the scenario.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-get_scenario_openness"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-get_scenario_openness}{}}}
\subsection{\code{DiseasyActivity$get_scenario_openness()}}{
  Return openness [0 ; 1] for all age groups and activities on all dates.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$get_scenario_openness(
  age_cuts_lower = NULL,
  population = NULL,
  weights = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used.}
      \item{\code{population}}{Output of \verb{$map_population()} to allow age mapping (will be removed soon).}
      \item{\code{weights}}{(\code{numeric(4)})\cr vector of weights for the four types of contacts. If \code{NULL}, no weighting is done. The weights are normalized before applying.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    (\code{list()})\cr
Returns a list with depth of two: value[[date]][[type]]
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-get_scenario_contacts"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-get_scenario_contacts}{}}}
\subsection{\code{DiseasyActivity$get_scenario_contacts()}}{
  Return contacts across age groups and activities on all dates.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$get_scenario_contacts(
  age_cuts_lower = NULL,
  population = NULL,
  weights = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used.}
      \item{\code{population}}{Output of \verb{$map_population()} to allow age mapping (will be removed soon).}
      \item{\code{weights}}{(\code{numeric(4)})\cr vector of weights for the four types of contacts. If \code{NULL}, no weighting is done.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    If no weights are supplied, a \code{list()} of depth of two: value[[date]][[type]] is returned.
Rescale contact matrices to population contact rates
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-rescale_contacts_to_rates"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-rescale_contacts_to_rates}{}}}
\subsection{\code{DiseasyActivity$rescale_contacts_to_rates()}}{
  Re-scale from contacts to rates per individual to fractional population.
@details
If the contact matrix is \eqn{\beta_{i,j}} and the population is \eqn{N_j}, then
this function returns the rescaled elements \eqn{\left(sum_j N_j\right)  \beta_{i,j} / N_j}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$rescale_contacts_to_rates(input, population)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{input}}{(\verb{matrix array} or \code{list}(\verb{matrix array}))\cr
Contacts to be re-scaled.}
      \item{\code{population}}{(\code{numeric})\cr
Population vector to weight contacts by.
Must use same age_groups as the contact matrix input.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Returns an object with the same structure as the input
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-reset_scenario"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-reset_scenario}{}}}
\subsection{\code{DiseasyActivity$reset_scenario()}}{
  Resets the scenario in the module.
NOTE: Called automatically when setting/changing activity units.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$reset_scenario()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-plot"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-plot}{}}}
\subsection{\code{DiseasyActivity$plot()}}{
  Plot the first set of contact matrices of the scenario as well as the "openness" over time.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$plot(
  age_cuts_lower = NULL,
  weights = rep(1, 4),
  contacts_date = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used.}
      \item{\code{weights}}{(\code{numeric(4)})\cr vector of weights for the four types of contacts. If \code{NULL}, no weighting is done.}
      \item{\code{contacts_date}}{(\code{Date(1)})\cr
The date to plot contact matrix for (default is earliest contact matrix).}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-describe"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-describe}{}}}
\subsection{\code{DiseasyActivity$describe()}}{
  Prints a human readable report of the internal state of the module.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$describe()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-clone}{}}}
\subsection{\code{DiseasyActivity$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyActivity$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
