% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyActivity.R
\name{DiseasyActivity}
\alias{DiseasyActivity}
\title{Activity handler}
\description{
TODO
}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasyActivity}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{scenario_matrix}}{(\code{matrix})\cr
A reduced view of the internal state of restrictions (showing only used activity units). Read-only.}

\item{\code{risk_matrix}}{(\code{matrix})\cr
A reduced view of the internal state of overall risk multipliers. Read-only.}

\item{\code{contact_basis}}{(\code{matrix})\cr
The basis of contacts between age groups used to compute activity. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyActivity-new}{\code{DiseasyActivity$new()}}
\item \href{#method-DiseasyActivity-set_activity_units}{\code{DiseasyActivity$set_activity_units()}}
\item \href{#method-DiseasyActivity-set_contact_basis}{\code{DiseasyActivity$set_contact_basis()}}
\item \href{#method-DiseasyActivity-change_activity}{\code{DiseasyActivity$change_activity()}}
\item \href{#method-DiseasyActivity-change_risk}{\code{DiseasyActivity$change_risk()}}
\item \href{#method-DiseasyActivity-crop_scenario}{\code{DiseasyActivity$crop_scenario()}}
\item \href{#method-DiseasyActivity-get_scenario_activities}{\code{DiseasyActivity$get_scenario_activities()}}
\item \href{#method-DiseasyActivity-get_scenario_openness}{\code{DiseasyActivity$get_scenario_openness()}}
\item \href{#method-DiseasyActivity-get_scenario_contacts}{\code{DiseasyActivity$get_scenario_contacts()}}
\item \href{#method-DiseasyActivity-rescale_counts2rates}{\code{DiseasyActivity$rescale_counts2rates()}}
\item \href{#method-DiseasyActivity-reset_scenario}{\code{DiseasyActivity$reset_scenario()}}
\item \href{#method-DiseasyActivity-describe}{\code{DiseasyActivity$describe()}}
\item \href{#method-DiseasyActivity-clone}{\code{DiseasyActivity$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="finalize"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-finalize'><code>diseasy::DiseasyBaseModule$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyActivity} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$new(
  base_scenario = "closed",
  activity_units = NULL,
  contact_basis = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_scenario}}{(\code{character(1)})\cr
Baseline scenario. Must be either fully "open" or "closed" or "dk_reference"}

\item{\code{activity_units}}{(\code{list}s of \code{list}s)\cr
A nested list of all possible "units" of activity that can be opened or closed}

\item{\code{contact_basis}}{(\code{list()})\cr
TODO}

\item{\code{...}}{Parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-set_activity_units"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-set_activity_units}{}}}
\subsection{Method \code{set_activity_units()}}{
Sets the list of all possible "units" of activity that can be opened or closed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$set_activity_units(activity_units)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{activity_units}}{(\code{list(list())})\cr
A nested list of all possible "units" of activity that can be opened or closed}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Each element in the activity_units list should be a list with the following elements: #TODO
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-set_contact_basis"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-set_contact_basis}{}}}
\subsection{Method \code{set_contact_basis()}}{
Sets the contact matrix basis the activity is computed from
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$set_contact_basis(contact_basis)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{contact_basis}}{(\code{list(list())})\cr
A nested list with all the needed information for the contact_basis\cr
\itemize{
\item \code{counts} contains the age stratified contact counts across the arenas of the basis
(e.g. "work", "home", "school", "other")\cr
\item \code{prop} #TODO\cr
\item \code{pop_ref_1yr} contains a \code{data.frame} with the columns\cr
\itemize{
\item \code{age} (\code{integer()}) 1-year age group\cr
\item \code{pop} (\code{numeric()}) size of population in age group\cr
\item \code{prop} (\code{numeric()}) proportion of total population in age group\cr
}
\item \code{description} contains information about the source of the contact basis
}}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
#TODO
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-change_activity"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-change_activity}{}}}
\subsection{Method \code{change_activity()}}{
Adds the specified openings and closings to the scenario
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$change_activity(date, opening = NA, closing = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{(\code{Date()} | \code{data.frame})\cr
Either a vector of dates when activity changes or a data.frame with columns 'date', 'opening' and 'closing'}

\item{\code{opening}}{(\code{character()})\cr
Names of activities to open on given date.
Omitted if \code{data.frame} is given to date argument}

\item{\code{closing}}{(\code{character()})\cr
Names of activities to close on given date.
Omitted if \code{data.frame} is given to date argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-change_risk"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-change_risk}{}}}
\subsection{Method \code{change_risk()}}{
Sets the overall risk of types of activity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$change_risk(date, type = NA, risk = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{Vector of dates when risk changes.
The first argument can also be a data.frame with columns "date", "type" and "risk"}

\item{\code{type}}{(\code{character})\cr
Name of activity type to change. Must be in "work", "school", "home" and "other"}

\item{\code{risk}}{Relative risk for the given type from the given date}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-crop_scenario"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-crop_scenario}{}}}
\subsection{Method \code{crop_scenario()}}{
Helper function to crop the scenario matrix in time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$crop_scenario(first_date = NULL, last_date = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{first_date}}{(\code{Date})\cr
New first date in scenario. The column for the lasted prior date will be new first column}

\item{\code{last_date}}{(\code{Date})\cr
All columns after this date are deleted}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-get_scenario_activities"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-get_scenario_activities}{}}}
\subsection{Method \code{get_scenario_activities()}}{
Return \code{list} with opened/closed activities on dates where there are changes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$get_scenario_activities()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with opened/closed activities on dates where there are changes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-get_scenario_openness"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-get_scenario_openness}{}}}
\subsection{Method \code{get_scenario_openness()}}{
Return openness [0 ; 1] for all age groups and activities on all dates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$get_scenario_openness(age_cuts_lower = NULL, weights = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{age_cuts_lower}}{(\code{numeric})\cr vector of ages defining the lower bound for each age group. If NULL (default), age groups of contact_basis is used.}

\item{\code{weights}}{(\code{numeric(4)})\cr vector of weights for the four types of contacts. If NULL (default), no weighting is done.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a list with depth of two: value[\link{date}][\link{type}]
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-get_scenario_contacts"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-get_scenario_contacts}{}}}
\subsection{Method \code{get_scenario_contacts()}}{
Return contacts for across age groups and activities on all dates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$get_scenario_contacts(age_cuts_lower = NULL, weights = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{age_cuts_lower}}{(\code{numeric})\cr vector of ages defining the lower bound for each age group. If NULL (default), age groups of contact_basis is used.}

\item{\code{weights}}{(\code{numeric(4)})\cr vector of weights for the four types of contacts. If NULL (default), no weighting is done.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If no weights are supplied, a \code{list()} of depth of two: value[\link{date}][\link{type}] is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-rescale_counts2rates"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-rescale_counts2rates}{}}}
\subsection{Method \code{rescale_counts2rates()}}{
Rescale from contacts to rates per individual to fractional population.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$rescale_counts2rates(input, pop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Matrix or (nested) list(s) to be rescaled.}

\item{\code{pop}}{Population vector. Is normalized within.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns an object with the same structure as the input
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-reset_scenario"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-reset_scenario}{}}}
\subsection{Method \code{reset_scenario()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$reset_scenario()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-describe"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-describe}{}}}
\subsection{Method \code{describe()}}{
Prints a human readable report of the internal state of the module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyActivity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyActivity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyActivity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
