% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyEnsemble.R
\name{DiseasyEnsemble-generics}
\alias{DiseasyEnsemble-generics}
\alias{print.DiseasyEnsemble}
\alias{summary.DiseasyEnsemble}
\alias{plot.DiseasyEnsemble}
\title{Standard generics for \code{DiseasyEnsemble} objects}
\usage{
\method{print}{DiseasyEnsemble}(x, width = 200, ...)

\method{summary}{DiseasyEnsemble}(object, ...)

\method{plot}{DiseasyEnsemble}(
  x,
  observable,
  prediction_length,
  stratification = NULL,
  context_length = 30,
  by_model = FALSE
)
}
\arguments{
\item{x, object}{(\code{DiseasyEnsemble})\cr
Ensemble object to print, summarise or plot.}

\item{width}{(\code{integer(1)})\cr
The maximum number of characters to print.}

\item{...}{additional arguments affecting the summary produced.}

\item{observable}{(\code{character})\cr The observable to provide data or prediction for.}

\item{prediction_length}{(\code{numeric})\cr The number of days to predict. The prediction start is defined by \code{last_queryable_date} of the \code{?DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{stratification}{(\code{list}(\code{quosures}) or \code{NULL})\cr Use \code{rlang::quos(...)} to specify stratification. If given, expressions in stratification evaluated to give the stratification level.}

\item{context_length}{(\code{integer(1)})\cr
Number of days prior to prediction to plot observable for.}

\item{by_model}{(\code{logical(1)})\cr
Should the plot be stratified by model?}
}
\value{
\code{NULL} (called for side effects)
}
\description{
Standard generics for \code{DiseasyEnsemble} objects
}
\examples{
\dontshow{if (rlang::is_installed("duckdb")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  observables <- DiseasyObservables$new(
    diseasystore = DiseasystoreSeirExample,
    conn = DBI::dbConnect(duckdb::duckdb())
  )

  # Create a DiseasyEnsemble object
  ensemble <- combineasy(
    model_templates = list(DiseasyModelG0, DiseasyModelG1)
    modules = observables
  )

  print(ensemble)

  summary(ensemble)

  plot(ensemble)

  rm(ensemble, observables)
\dontshow{\}) # examplesIf}
}
