% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyImmunity.R
\name{DiseasyImmunity}
\alias{DiseasyImmunity}
\title{Diseasy' immunity handler}
\description{
The \code{DiseasyImmunity} module is responsible for implementing various models for the immunity dependency of the
diseases. Meaning the different scenarios there can be of waning of immunity.
The module implements a number immunity models with different functional forms.

See the vignette("diseasy-immunity") for examples of use.
@return
A new instance of the \code{DiseasyImmunity} \link[R6:R6Class]{R6} class.
}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasyImmunity}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{available_waning_models}}{(\code{character})\cr
The list of available waning models}

\item{\code{model}}{(\code{function})\cr
The model currently being used in the module. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyImmunity-new}{\code{DiseasyImmunity$new()}}
\item \href{#method-DiseasyImmunity-set_time_scale}{\code{DiseasyImmunity$set_time_scale()}}
\item \href{#method-DiseasyImmunity-use_waning_model}{\code{DiseasyImmunity$use_waning_model()}}
\item \href{#method-DiseasyImmunity-use_no_waning}{\code{DiseasyImmunity$use_no_waning()}}
\item \href{#method-DiseasyImmunity-use_exponential_waning}{\code{DiseasyImmunity$use_exponential_waning()}}
\item \href{#method-DiseasyImmunity-use_sigmoidal_waning}{\code{DiseasyImmunity$use_sigmoidal_waning()}}
\item \href{#method-DiseasyImmunity-use_linear_waning}{\code{DiseasyImmunity$use_linear_waning()}}
\item \href{#method-DiseasyImmunity-use_heaviside_waning}{\code{DiseasyImmunity$use_heaviside_waning()}}
\item \href{#method-DiseasyImmunity-clone}{\code{DiseasyImmunity$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="finalize"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-finalize'><code>diseasy::DiseasyBaseModule$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyImmunity} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-set_time_scale"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-set_time_scale}{}}}
\subsection{Method \code{set_time_scale()}}{
Sets the time scale for the waning of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$set_time_scale(time_scale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scale}}{(\code{numeric})\cr
The time_scale determines the rate of decay in the model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-use_waning_model"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-use_waning_model}{}}}
\subsection{Method \code{use_waning_model()}}{
Sets the \code{DiseasyImmunity} module to use the specified waning model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$use_waning_model(model_name, dots = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{(\code{character})\cr
Name of the waning_model to use (calls the equivalent $use_<model_name>()).}

\item{\code{dots}}{(\code{list})\cr
Named list of arguments that will be passed at dot-ellipsis to the waning model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-use_no_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-use_no_waning}{}}}
\subsection{Method \code{use_no_waning()}}{
Retrieves the waning model with a constant value (1).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$use_no_waning()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-use_exponential_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-use_exponential_waning}{}}}
\subsection{Method \code{use_exponential_waning()}}{
Sets the \code{DiseasyImmunity} module to use an exponential model for waning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$use_exponential_waning(time_scale = 20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scale}}{\verb{r rd_time_scale()}
By default, it is set to 20}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-use_sigmoidal_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-use_sigmoidal_waning}{}}}
\subsection{Method \code{use_sigmoidal_waning()}}{
Sets the \code{DiseasyImmunity} module to use a sigmoidal model for waning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$use_sigmoidal_waning(time_scale = 20, shape = 6)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scale}}{\verb{r rd_time_scale()}
By default, it is set to 20}

\item{\code{shape}}{(\code{numeric})\cr
Determines the steepness of the waning curve in the sigmoidal waning model.
Higher values of \code{shape} result in a steeper curve, leading to a more rapid decline in immunity.
By default, it is set to 6.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-use_linear_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-use_linear_waning}{}}}
\subsection{Method \code{use_linear_waning()}}{
Sets the \code{DiseasyImmunity} module to use a linear model for waning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$use_linear_waning(time_scale = 20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scale}}{\verb{r rd_time_scale()}
By default, it is set to 20}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-use_heaviside_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-use_heaviside_waning}{}}}
\subsection{Method \code{use_heaviside_waning()}}{
Sets the \code{DiseasyImmunity} module to use a heaviside model for waning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$use_heaviside_waning(time_scale = 20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scale}}{\verb{r rd_time_scale()}
By default, it is set to 20}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
