% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyImmunity.R
\name{DiseasyImmunity}
\alias{DiseasyImmunity}
\title{Diseasy' immunity handler}
\description{
The \code{DiseasyImmunity} module is responsible for implementing various models for the immunity dependency of the
diseases. Meaning the different scenarios there can be of waning of immunity.
The module implements a number immunity models with different functional forms.

See the vignette("diseasy-immunity") for examples of use.
@return
A new instance of the \code{DiseasyImmunity} \link[R6:R6Class]{R6} class.
}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasyImmunity}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rate}}{(\code{list})\cr
The list of rates created by the selected approach. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{available_waning_models}}{(\code{character})\cr
The list of available waning models}

\item{\code{model}}{(\code{function})\cr
The model currently being used in the module. Read-only.}

\item{\code{rate}}{(\code{list})\cr
The list of rates created by the selected approach. Read-only.}

\item{\code{available_approaches}}{(\code{character})\cr
The list of available approaches}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyImmunity-new}{\code{DiseasyImmunity$new()}}
\item \href{#method-DiseasyImmunity-set_time_scales}{\code{DiseasyImmunity$set_time_scales()}}
\item \href{#method-DiseasyImmunity-set_waning_model}{\code{DiseasyImmunity$set_waning_model()}}
\item \href{#method-DiseasyImmunity-set_no_waning}{\code{DiseasyImmunity$set_no_waning()}}
\item \href{#method-DiseasyImmunity-set_exponential_waning}{\code{DiseasyImmunity$set_exponential_waning()}}
\item \href{#method-DiseasyImmunity-set_sigmoidal_waning}{\code{DiseasyImmunity$set_sigmoidal_waning()}}
\item \href{#method-DiseasyImmunity-set_linear_waning}{\code{DiseasyImmunity$set_linear_waning()}}
\item \href{#method-DiseasyImmunity-set_heaviside_waning}{\code{DiseasyImmunity$set_heaviside_waning()}}
\item \href{#method-DiseasyImmunity-set_custom_waning}{\code{DiseasyImmunity$set_custom_waning()}}
\item \href{#method-DiseasyImmunity-approximate_compartmental}{\code{DiseasyImmunity$approximate_compartmental()}}
\item \href{#method-DiseasyImmunity-clone}{\code{DiseasyImmunity$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="finalize"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-finalize'><code>diseasy::DiseasyBaseModule$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyImmunity} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-set_time_scales"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-set_time_scales}{}}}
\subsection{Method \code{set_time_scales()}}{
Sets the characteristic time scale for the waning of the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$set_time_scales(time_scales = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scales}}{(\code{list})\cr
A named list of target and new time_scale (can be multiple)
An exmaple could be; list(infection = 60)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-set_waning_model"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-set_waning_model}{}}}
\subsection{Method \code{set_waning_model()}}{
Sets the \code{DiseasyImmunity} module to use the specified waning model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$set_waning_model(model_name, dots = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{(\code{character})\cr
Name of the waning_model to use (calls the equivalent $use_<model_name>()).}

\item{\code{dots}}{(\code{list})\cr
Named list of arguments that will be passed at dot-ellipsis to the waning model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-set_no_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-set_no_waning}{}}}
\subsection{Method \code{set_no_waning()}}{
Retrieves the waning model with a constant value (1).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$set_no_waning(target = "infection")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target}}{(\code{character})\cr
The target of the waning model (f.x. infection, hospitalisation, death).
By default, it is set to "infection".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-set_exponential_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-set_exponential_waning}{}}}
\subsection{Method \code{set_exponential_waning()}}{
Sets the \code{DiseasyImmunity} module to set an exponential model for waning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$set_exponential_waning(time_scale = 20, target = "infection")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scale}}{(\code{numeric})\cr Sets the time_scale of the waning (immunity) model. The time_scale is the number of days before half of the immunity has waned
By default, it is set to 20}

\item{\code{target}}{(\code{character})\cr
The target of the waning model (f.x. infection, hospitalisation, death).
By default, it is set to "infection".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-set_sigmoidal_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-set_sigmoidal_waning}{}}}
\subsection{Method \code{set_sigmoidal_waning()}}{
Sets the \code{DiseasyImmunity} module to set a sigmoidal model for waning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$set_sigmoidal_waning(
  time_scale = 20,
  shape = 6,
  target = "infection"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scale}}{(\code{numeric})\cr Sets the time_scale of the waning (immunity) model. The time_scale is the number of days before half of the immunity has waned
By default, it is set to 20}

\item{\code{shape}}{(\code{numeric})\cr
Determines the steepness of the waning curve in the sigmoidal waning model.
Higher values of \code{shape} result in a steeper curve, leading to a more rapid decline in immunity.
By default, it is set to 6.}

\item{\code{target}}{(\code{character})\cr
The target of the waning model (f.x. infection, hospitalisation, death).
By default, it is set to "infection".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-set_linear_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-set_linear_waning}{}}}
\subsection{Method \code{set_linear_waning()}}{
Sets the \code{DiseasyImmunity} module to set a linear model for waning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$set_linear_waning(time_scale = 20, target = "infection")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scale}}{(\code{numeric})\cr Sets the time_scale of the waning (immunity) model. The time_scale is the number of days before half of the immunity has waned
By default, it is set to 20}

\item{\code{target}}{(\code{character})\cr
The target of the waning model (f.x. infection, hospitalisation, death).
By default, it is set to "infection".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-set_heaviside_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-set_heaviside_waning}{}}}
\subsection{Method \code{set_heaviside_waning()}}{
Sets the \code{DiseasyImmunity} module to set a heaviside model for waning.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$set_heaviside_waning(time_scale = 20, target = "infection")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_scale}}{(\code{numeric})\cr Sets the time_scale of the waning (immunity) model. The time_scale is the number of days before half of the immunity has waned
By default, it is set to 20}

\item{\code{target}}{(\code{character})\cr
The target of the waning model (f.x. infection, hospitalisation, death).
By default, it is set to "infection".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-set_custom_waning"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-set_custom_waning}{}}}
\subsection{Method \code{set_custom_waning()}}{
Sets the \code{DiseasyImmunity} module to set a custom waning function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$set_custom_waning(
  custom_function = function(t) 1/(1 + exp((t - time_scale)/time_scale)),
  time_scale = 20,
  target = "infection",
  name = "custom_waning"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{custom_function}}{(\code{function})\cr
Set a custom waning function in following format: \(t)}

\item{\code{time_scale}}{(\code{numeric})\cr Sets the time_scale of the waning (immunity) model. The time_scale is the number of days before half of the immunity has waned
By default, it is set to 20}

\item{\code{target}}{(\code{character})\cr
The target of the waning model (f.x. infection, hospitalisation, death).
By default, it is set to "infection".}

\item{\code{name}}{Set the name of the custom waning function
By default, it is set to "custom_waning"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-approximate_compartmental"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-approximate_compartmental}{}}}
\subsection{Method \code{approximate_compartmental()}}{
The function creates N compartments for the models and does a phase-type distribution based on the
time scales, resulting in the probability to be in each compartment.
The models are then optimized to find the best fitting rates going out and between the compartments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$approximate_compartmental(
  approach = c("rate_equal", "gamma_fixed_step", "all_free"),
  N = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{approach}}{(\code{str} or \code{numeric})\cr
Specifies the approach to be used from the available approaches.
It can be provided as a string with the approach name "rate_equal", "gamma_fixed_step" or "all_free".
or as a numeric value representing the approach index 1, 2, or 3.}

\item{\code{N}}{(\code{numeric})\cr
Number of compartments to be used in the model.
By default, it is set to 5}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyImmunity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyImmunity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyImmunity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
