% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyModelOde.R
\name{DiseasyModelOde}
\alias{DiseasyModelOde}
\title{Meta module for the regression class of models}
\value{
A new instance of the \code{DiseasyModelOde} \link[R6:R6Class]{R6} class.
}
\description{
TODO
}
\examples{
  # Normally, you would not want to create this module directly, but it is possible.
  module <- DiseasyModelOde$new()

  rm(module)
}
\seealso{
\link{DiseasyObservables}
}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{diseasy::DiseasyModel}} -> \code{DiseasyModelOde}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyModelOde-get_results}{\code{DiseasyModelOde$get_results()}}
\item \href{#method-DiseasyModelOde-clone}{\code{DiseasyModelOde$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="finalize"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-finalize'><code>diseasy::DiseasyBaseModule$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_data'><code>diseasy::DiseasyModel$get_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="initialize"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-initialize'><code>diseasy::DiseasyModel$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOde-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOde-get_results}{}}}
\subsection{Method \code{get_results()}}{
The primary method used to request model results of a given observable at a given stratification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOde$get_results(
  observable,
  prediction_length,
  quantiles = NULL,
  stratification = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{observable}}{(\code{character})\cr The observable to provide data or prediction for.}

\item{\code{prediction_length}}{(\code{numeric})\cr The number of days to predict. The prediction start is defined by \code{last_queryable_date} of the \code{?DiseasyObservables} \link[R6:R6Class]{R6} class.}

\item{\code{quantiles}}{(\code{list}(\code{numeric}))\cr Default NULL. If given, results are returned at the quantiles given.}

\item{\code{stratification}}{(\code{list}(\code{quosures}))\cr Default NULL. Use \code{rlang::quos(...)} to specify stratification. If given, expressions in stratification evaluated to give the stratification level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble} \link[tibble:tibble]{tibble::tibble} with predictions at the level specified by stratification level. In addition to stratification columns, the output has the columns:\cr   date (\code{Date}) specifying the date of the prediction\cr   realization_id (\code{character}) giving a unique id for each realization in the ensemble\cr   model (\code{character}) the name (classname) of the model used to provide the prediction.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOde-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOde-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOde$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
