% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyModelOdeSeir.R
\name{DiseasyModelOdeSeir}
\alias{DiseasyModelOdeSeir}
\title{TODO}
\description{
TODO
}
\keyword{model-template}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{diseasy::DiseasyModel}} -> \code{DiseasyModelOdeSeir}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{immunity}}{Placeholder for the immunity module}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyModelOdeSeir-new}{\code{DiseasyModelOdeSeir$new()}}
\item \href{#method-DiseasyModelOdeSeir-set_forcing_functions}{\code{DiseasyModelOdeSeir$set_forcing_functions()}}
\item \href{#method-DiseasyModelOdeSeir-clone}{\code{DiseasyModelOdeSeir$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="finalize"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-finalize'><code>diseasy::DiseasyBaseModule$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_results"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_results'><code>diseasy::DiseasyModel$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_training_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_training_data'><code>diseasy::DiseasyModel$get_training_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOdeSeir-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOdeSeir-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyModelOdeSeir} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOdeSeir$new(
  compartment_structure = c(E = 2L, I = 3L, R = 2L),
  disease_progression_rates = c(E = 1, I = 1),
  malthusian_matching = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{compartment_structure}}{(\verb{named integer()})\cr
The structure of the compartments in the model.
The names should be \code{E}, \code{I}, and \code{R} for the exposed, infectious, and recovered compartments, respectively.
The exposed compartments can optionally be omitted.}

\item{\code{disease_progression_rates}}{(\verb{named numeric()})\cr
The overall progression rates for the disease states.
The reciprocal of each rate is the average time spent in the all of the corresponding compartments.
The exposed compartments can optionally be omitted.}

\item{\code{malthusian_matching}}{(\code{logical(1)})\cr
Should the model be scaled such the Malthusian growth rate marches the corresponding SIR model?}

\item{\code{...}}{parameters sent to \code{DiseasyModel} \link[R6:R6Class]{R6} constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOdeSeir-set_forcing_functions"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOdeSeir-set_forcing_functions}{}}}
\subsection{Method \code{set_forcing_functions()}}{
Set the forcing functions for the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOdeSeir$set_forcing_functions(
  infected_forcing = NULL,
  state_vector_forcing = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{infected_forcing}}{(\code{function})\cr
A function that takes arguments \code{t} and \code{infected} and modifies the number of infected at time \code{t}.}

\item{\code{state_vector_forcing}}{(\code{function})\cr
A function that takes arguments \code{t}, \code{dy_dt} and \code{new_infections} and modifies the flow into the
compartments at time \code{t}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOdeSeir-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOdeSeir-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOdeSeir$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
