% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyModelOdeSeir.R
\name{DiseasyModelOdeSeir}
\alias{DiseasyModelOdeSeir}
\title{A flexible SEIR model template}
\value{
A new instance of the \code{DiseasyModelOdeSeir} \link[R6:R6Class]{R6} class.
}
\description{
This class provides a \code{diseasy} model-template for a compartmental SEIR ordinary differential equation model.
The number of consecutive exposed, infectious, and recovered compartments can flexibly be specified to generate a
number of structurally different SEIR models.

Similarly, the number of age groups in the model can also be controlled to create structurally different models.

The model implements the following features:
\itemize{
\item A scaling of infection risk based on season (via \code{DiseasySeason})
\item Contact matrices and activity scenarios (via \code{DiseasyActivity})
\item Waning of immunity (via \code{DiseasyImmunity})
\item Asymmetric cross-immunity interactions between variants (via \code{DiseasyVariant})
}

See \code{vignette(diseasy-model-ode-seir)} for a detailed examples of how to use this model.
}
\examples{
\dontshow{if (rlang::is_installed("RSQLite")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  The model can be instantiated almost without arguments, but a observables modules needs to be specified.
  obs <- DiseasyObservables$new(diseasystore = "Google COVID-19",
                                conn = DBI::dbConnect(RSQLite::SQLite()))

  # We create a default instance which has:
  # * 1 age group (0+)
  # * 1 variant
  # * No season scaling
  # * No activity scenarios
  m <- DiseasyModelOdeSeir$new(observables = obs)

  # TODO: Continue this minimal example once module is complete

  rm(m)
\dontshow{\}) # examplesIf}
}
\keyword{model-template}
\section{Super classes}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{\link[diseasy:DiseasyModel]{diseasy::DiseasyModel}} -> \code{DiseasyModelOdeSeir}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{immunity}}{Place holder for the immunity module}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{compartment_structure}}{(\verb{named integer()})\cr The structure of the compartments in the model. The names should be \code{E}, \code{I}, and \code{R} for the exposed, infectious, and recovered compartments, respectively.  Read only.}

\item{\code{disease_progression_rates}}{(\verb{named numeric()})\cr The overall progression rates for the disease states. The reciprocal of each rate is the average time spent in the all of the corresponding compartments.  Read only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyModelOdeSeir-new}{\code{DiseasyModelOdeSeir$new()}}
\item \href{#method-DiseasyModelOdeSeir-set_forcing_functions}{\code{DiseasyModelOdeSeir$set_forcing_functions()}}
\item \href{#method-DiseasyModelOdeSeir-malthusian_growth_rate}{\code{DiseasyModelOdeSeir$malthusian_growth_rate()}}
\item \href{#method-DiseasyModelOdeSeir-initialise_state_vector}{\code{DiseasyModelOdeSeir$initialise_state_vector()}}
\item \href{#method-DiseasyModelOdeSeir-rhs}{\code{DiseasyModelOdeSeir$rhs()}}
\item \href{#method-DiseasyModelOdeSeir-clone}{\code{DiseasyModelOdeSeir$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="finalize"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-finalize'><code>diseasy::DiseasyBaseModule$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_results"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_results'><code>diseasy::DiseasyModel$get_results()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyModel" data-id="get_training_data"><a href='../../diseasy/html/DiseasyModel.html#method-DiseasyModel-get_training_data'><code>diseasy::DiseasyModel$get_training_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOdeSeir-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOdeSeir-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the \code{DiseasyModelOdeSeir} \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOdeSeir$new(
  compartment_structure = c(E = 2L, I = 3L, R = 2L),
  disease_progression_rates = c(E = 1, I = 1),
  malthusian_matching = TRUE,
  activity = TRUE,
  season = TRUE,
  variant = TRUE,
  parameters = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{compartment_structure}}{(\verb{named integer()})\cr
The structure of the compartments in the model.
The names should be \code{E}, \code{I}, and \code{R} for the exposed, infectious, and recovered compartments, respectively.
The exposed compartments can optionally be omitted.}

\item{\code{disease_progression_rates}}{(\verb{named numeric()})\cr
The overall progression rates for the disease states.
The reciprocal of each rate is the average time spent in the all of the corresponding compartments.
The exposed compartments can optionally be omitted.}

\item{\code{malthusian_matching}}{(\code{logical(1)})\cr
Should the model be scaled such the Malthusian growth rate marches the corresponding SIR model?}

\item{\code{activity, season, variant}}{(\code{boolean} or \verb{R6::R6Class instance})\cr If a boolean is given, it dictates whether to load a new instance module of this class.\cr If an instance of the module is provided instead, a copy of this instance is added to the \code{DiseasyModel} instance. This copy is a "clone" of the instance at the time it is added and any subsequent changes to the instance will not reflect in the copy that is added to \code{DiseasyModel}.}

\item{\code{parameters}}{(\verb{named list()})\cr
List of parameters to set for the model during initialization.

Parameters controlling the structure of the model:
\itemize{
\item \code{age_cuts_lower} - Determines the age groups in the model.
}

Parameters controlling the dynamics of the model:
(Can be inferred from observational data via initialisation routines)
\itemize{
\item \code{overall_infection_risk} - A scalar that scales contact rates to infection rates.
}

Parameters controlling initialisation routines
\itemize{
\item \code{incidence_polynomial_order} - The degree of the polynomial to fit to the incidence curves.
}

Parameters controlling the functional modules:
\itemize{
\item \code{activity.weights} - passed to \code{DiseasyActivity$get_scenario_contacts(..., weights = activity.weights)}
\item \code{immunity.method} - passed to \code{DiseasyImmunity$approximate_compartmental(method = immunity.method, ...)}
}

Additional parameters are:
\itemize{
\item \code{training_length} (\verb{named numeric})\cr   The number of days that should be included in the training splits of the data for the model.   Allowed splits are: "training", "testing", and "validation".
}}

\item{\code{...}}{Parameters sent to \code{DiseasyModel} \link[R6:R6Class]{R6} constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOdeSeir-set_forcing_functions"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOdeSeir-set_forcing_functions}{}}}
\subsection{Method \code{set_forcing_functions()}}{
Set the forcing functions for the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOdeSeir$set_forcing_functions(
  infected_forcing = NULL,
  state_vector_forcing = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{infected_forcing}}{(\code{function})\cr
A function that takes arguments \code{t} and \code{infected} and modifies the number of infected at time \code{t}.}

\item{\code{state_vector_forcing}}{(\code{function})\cr
A function that takes arguments \code{t} and\code{dy_dt} and modifies the flow into the
compartments at time \code{t}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL (called for side effects)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOdeSeir-malthusian_growth_rate"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOdeSeir-malthusian_growth_rate}{}}}
\subsection{Method \code{malthusian_growth_rate()}}{
This function computes the Malthusian growth rate for the given model configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOdeSeir$malthusian_growth_rate(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to \verb{$generator_matrix()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This section follows the method outlined in doi: 10.1098/rsif.2009.0386
To compute the scaling, we need to compute the Jacobian matrix of the linearised system.
Here, we linearise around S = 1.
In this limit, there is no interaction with variants (since everyone is susceptible).
The Malthusian growth rate is therefore not dependent on factors such as cross-immunity.
}

\subsection{Returns}{
(\code{numeric(1)})\cr
The Malthusian growth rate for the model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOdeSeir-initialise_state_vector"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOdeSeir-initialise_state_vector}{}}}
\subsection{Method \code{initialise_state_vector()}}{
Infer the state_vector from incidence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOdeSeir$initialise_state_vector(
  incidence_data,
  method = c("derivative", "eigen-value")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{incidence_data}}{(\code{data.frame})\cr
Incidence observations as a \code{data.frame} with columns
\itemize{
\item \code{date}: The date of the observations
\item \code{age_group}: The age group of the incidence observation (following \code{diseasystore::age_labels()} format)
\item \code{variant}: The variant of the incidence observation.
\item \code{incidence}: The incidence in the age group at the given date
}}

\item{\code{method}}{(\code{character(1)})\cr
The method to use for initialising the state vector.
\itemize{
\item \code{derivative}: Directly infers the EI compartments from derivatives of the incidence signal.
\item \code{eigen-value}: Uses the eigenvalues of the generator matrix to infer the state vector.
}

See the article \code{SEIR-initialisation} in the online documentation for more information.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{numeric()})\cr
The initial state vector for the model (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyModelOdeSeir-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyModelOdeSeir-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyModelOdeSeir$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
