% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyPopulation.R
\name{DiseasyPopulation}
\alias{DiseasyPopulation}
\title{Configure the model population}
\value{
A new instance of the \code{DiseasyPopulation} \link[R6:R6Class]{R6} class.
}
\description{
The \code{DiseasyPopulation} module is responsible for handling the population included in the model.

See vignette("diseasy-population").
}
\examples{
  # Create population module
  population <- DiseasyPopulation$new()

  # By default a single, non-stratified population is used in the models.
  population

  # Stratification can be added via methods

  # Stratifying by age
  population$stratify_age(age_cuts_lower = c(0, 60)) # 2 age groups
  # NB: Age cuts must be available in demography and disease data.

  population

  rm(population)
}
\seealso{
\link{DiseasyActivity}

\link{DiseasyRegions}
}
\keyword{functional-module}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{DiseasyBaseModule}} -> \code{DiseasyPopulation}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{groups}}{(\code{data.frame()})\cr
The demographic groups that have been configured in the module.}

    \item{\code{model_population}}{(\code{tibble})\cr
The population groups and their sizes configured in the module.}

    \item{\code{regional_stratification}}{(\code{character()})\cr The level of spatial stratification of the population. Read only.}

    \item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used. Read only.}

    \item{\code{activity}}{(\code{diseasy::DiseasyActivity})\cr
The local copy of an DiseasyActivity module. Read-only.}

    \item{\code{regions}}{(\code{diseasy::DiseasyRegions})\cr
The local copy of an DiseasyRegions module. Read-only.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DiseasyPopulation-initialize}{\code{DiseasyPopulation$new()}}
    \item \href{#method-DiseasyPopulation-stratify_age}{\code{DiseasyPopulation$stratify_age()}}
    \item \href{#method-DiseasyPopulation-stratify_regions}{\code{DiseasyPopulation$stratify_regions()}}
    \item \href{#method-DiseasyPopulation-validate_configuration}{\code{DiseasyPopulation$validate_configuration()}}
    \item \href{#method-DiseasyPopulation-per_capita_contact_matrices}{\code{DiseasyPopulation$per_capita_contact_matrices()}}
    \item \href{#method-DiseasyPopulation-map_population}{\code{DiseasyPopulation$map_population()}}
    \item \href{#method-DiseasyPopulation-describe}{\code{DiseasyPopulation$describe()}}
    \item \href{#method-DiseasyPopulation-clone}{\code{DiseasyPopulation$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>DiseasyBaseModule$load_module()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-initialize}{}}}
\subsection{\code{DiseasyPopulation$new()}}{
  Creates a new instance of the \code{DiseasyPopulation} \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyPopulation$new(
  age_cuts_lower = 0L,
  regional_stratification = NULL,
  regions = NULL,
  ...
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used.}
      \item{\code{regional_stratification}}{(\code{character()})\cr The level of spatial stratification of the population.}
      \item{\code{regions}}{(\code{DiseasyRegions})\cr
An instance of a regional module which should provide the demography of the population.}
      \item{\code{...}}{Parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-stratify_age"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-stratify_age}{}}}
\subsection{\code{DiseasyPopulation$stratify_age()}}{
  Sets the age stratification of the model population.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyPopulation$stratify_age(age_cuts_lower)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-stratify_regions"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-stratify_regions}{}}}
\subsection{\code{DiseasyPopulation$stratify_regions()}}{
  Sets the spatial stratification of the model population.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyPopulation$stratify_regions(regional_stratification)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{regional_stratification}}{(\code{character()})\cr The level of spatial stratification of the population.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-validate_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-validate_configuration}{}}}
\subsection{\code{DiseasyPopulation$validate_configuration()}}{
  Validate the stratifications are supported by demography data.
Throw error if misconfigured.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyPopulation$validate_configuration(
  age_cuts_lower = self \%.\% age_cuts_lower,
  regional_stratification = self \%.\% regional_stratification
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used.}
      \item{\code{regional_stratification}}{(\code{character()})\cr The level of spatial stratification of the population.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{NULL} (called for side effects)
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-per_capita_contact_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-per_capita_contact_matrices}{}}}
\subsection{\code{DiseasyPopulation$per_capita_contact_matrices()}}{
  Compute the per-capita contact matrices. See vignette("diseasy-activity") for details.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyPopulation$per_capita_contact_matrices(weights = rep(1, 4))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{weights}}{(\code{numeric(4)})\cr vector of weights for the four types of contacts. If \code{NULL}, no weighting is done.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{list}(\code{matrix})\cr
A \code{list} (with names indicating the dates of changes in contacts)
of contact rates (\code{matrix}).
Map population between age groups
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-map_population"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-map_population}{}}}
\subsection{\code{DiseasyPopulation$map_population()}}{
  The function computes the proportion of population in the new and old age groups.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyPopulation$map_population(
  age_cuts_lower = self \%.\% age_cuts_lower,
  regions = purrr::pluck(self \%.\% groups, "region", unique),
  age_groups_reference = NULL,
  demography = self \%.\% regions \%.\% demography
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used.}
      \item{\code{regions}}{(\code{character()})\cr
The regions the population should be mapped to (e.g. NUTS 1 regions).}
      \item{\code{age_groups_reference}}{(\code{character()})\cr
Age labels (created by \code{diseasystore::age_labels()} of reference data.}
      \item{\code{demography}}{(\code{data.frame})\cr
"A \code{data.frame} with the columns\\cr",
"  * \code{age} (\code{integer()}) 1-year age groups or \code{age_group} (\code{integer()}) dynamic age groups\\cr",
"  * \code{population} (\code{numeric()}) size of population in age group\\cr"}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \code{data.frame} which maps the age groups from their reference in \code{contact_basis} to
those supplied to the function.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-describe"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-describe}{}}}
\subsection{\code{DiseasyPopulation$describe()}}{
  Prints a human readable report of the internal state of the module.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyPopulation$describe()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-clone}{}}}
\subsection{\code{DiseasyPopulation$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyPopulation$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
