% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyPopulation.R
\name{DiseasyPopulation}
\alias{DiseasyPopulation}
\title{Configure the model population}
\value{
A new instance of the \code{DiseasyPopulation} \link[R6:R6Class]{R6} class.
}
\description{
The \code{DiseasyPopulation} module is responsible for handling the population included in the model.

See vignette("diseasy-population").
}
\examples{
  # Create population module
  population <- DiseasyPopulation$new()

  # By default a single, non-stratified population is used in the models.
  population

  # Stratification can be added via methods

  # Stratifying by age
  population$stratify_age(age_cuts_lower = c(0, 60)) # 2 age groups
  # NB: Age cuts must be available in demography and disease data.

  population

  rm(population)
}
\keyword{functional-module}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasyPopulation}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used. Read only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyPopulation-stratify_age}{\code{DiseasyPopulation$stratify_age()}}
\item \href{#method-DiseasyPopulation-describe}{\code{DiseasyPopulation$describe()}}
\item \href{#method-DiseasyPopulation-clone}{\code{DiseasyPopulation$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="initialize"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-initialize'><code>diseasy::DiseasyBaseModule$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-stratify_age"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-stratify_age}{}}}
\subsection{Method \code{stratify_age()}}{
Sets the age stratification of the model population.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyPopulation$stratify_age(age_cuts_lower)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{age_cuts_lower}}{(\code{integer()})\cr vector of ages defining the lower bound for each age group. If \code{NULL}, age groups of contact_basis is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-describe"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-describe}{}}}
\subsection{Method \code{describe()}}{
Prints a human readable report of the internal state of the module.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyPopulation$describe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyPopulation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyPopulation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyPopulation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
