% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyRegions.R
\name{DiseasyRegions}
\alias{DiseasyRegions}
\title{Diseasy' region (spatial) handler}
\value{
A new instance of the \code{DiseasyRegions} \link[R6:R6Class]{R6} class.
}
\description{
TODO

See the vignette("diseasy-region") for examples of use.
}
\examples{
  # TODO
}
\keyword{functional-module}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{DiseasyBaseModule}} -> \code{DiseasyRegions}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{regions}}{(\code{character()})\cr The geographic area of interest (expressed in NUTS codes).   Read only.}

    \item{\code{adjacency}}{(\code{data.frame(1)})\cr The adjacency (connectedness) of the regions (identified by NUTS codes). Effectively, the \code{adjacency} is a long-form of the adjacency-matrix with additional information about the spatial embedding of the regions. The \code{data.frame} must include the following columns:\cr - \code{from}  \cr - \code{to}  \cr - \code{adjacency}: strength of the connectedness (will be normalised).\cr    Read only.}

    \item{\code{demography}}{(\code{data.frame(1)})\cr The demography of the population per region. The \code{data.frame} must include the following columns:\cr * \code{nuts} (\code{character}) The NUTS 1/2/3 code for the stratification. Only one NUTS level is needed.\cr * \code{...} Optional additional stratification columns (e.g. sex, ethnicity).\cr * \code{population} (\code{numeric}) Number of individuals in the group. Read only.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DiseasyRegions-initialize}{\code{DiseasyRegions$new()}}
    \item \href{#method-DiseasyRegions-set_regions}{\code{DiseasyRegions$set_regions()}}
    \item \href{#method-DiseasyRegions-set_adjacency}{\code{DiseasyRegions$set_adjacency()}}
    \item \href{#method-DiseasyRegions-set_demography}{\code{DiseasyRegions$set_demography()}}
    \item \href{#method-DiseasyRegions-describe}{\code{DiseasyRegions$describe()}}
    \item \href{#method-DiseasyRegions-clone}{\code{DiseasyRegions$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>DiseasyBaseModule$load_module()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyRegions-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyRegions-initialize}{}}}
\subsection{\code{DiseasyRegions$new()}}{
  Creates a new instance of the \code{DiseasyRegions} \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyRegions$new(regions, adjacency, demography, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{regions}}{(\code{character()})\cr The geographic area of interest (expressed in NUTS codes).  The specified NUTS regions must be available in the \code{demography} and \code{adjacency} data.}
      \item{\code{adjacency}}{(\code{data.frame(1)})\cr The adjacency (connectedness) of the regions (identified by NUTS codes). Effectively, the \code{adjacency} is a long-form of the adjacency-matrix with additional information about the spatial embedding of the regions. The \code{data.frame} must include the following columns:\cr - \code{from} (NUTS 1/2/3 code *) \cr - \code{to} (NUTS 1/2/3 code *) \cr - \code{adjacency}: strength of the connectedness (will be normalised).\cr * only one NUTS level should be specified Only "upper" or "lower" triangle of the adjacency needs to be specified if the full matrix is specified, a symmetric variant is formed.}
      \item{\code{demography}}{(\code{data.frame(1)})\cr The demography of the population per region. The \code{data.frame} must include the following columns:\cr * \code{nuts} (\code{character}) The NUTS 1/2/3 code for the stratification. Only one NUTS level is needed.\cr * \code{...} Optional additional stratification columns (e.g. sex, ethnicity).\cr * \code{population} (\code{numeric}) Number of individuals in the group.}
      \item{\code{...}}{parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyRegions-set_regions"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyRegions-set_regions}{}}}
\subsection{\code{DiseasyRegions$set_regions()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyRegions$set_regions(regions)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{regions}}{(\code{character()})\cr The geographic area of interest (expressed in NUTS codes).  The specified NUTS regions must be available in the \code{demography} and \code{adjacency} data.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyRegions-set_adjacency"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyRegions-set_adjacency}{}}}
\subsection{\code{DiseasyRegions$set_adjacency()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyRegions$set_adjacency(adjacency)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{adjacency}}{(\code{data.frame(1)})\cr The adjacency (connectedness) of the regions (identified by NUTS codes). Effectively, the \code{adjacency} is a long-form of the adjacency-matrix with additional information about the spatial embedding of the regions. The \code{data.frame} must include the following columns:\cr - \code{from} (NUTS 1/2/3 code *) \cr - \code{to} (NUTS 1/2/3 code *) \cr - \code{adjacency}: strength of the connectedness (will be normalised).\cr * only one NUTS level should be specified Only "upper" or "lower" triangle of the adjacency needs to be specified if the full matrix is specified, a symmetric variant is formed.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyRegions-set_demography"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyRegions-set_demography}{}}}
\subsection{\code{DiseasyRegions$set_demography()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyRegions$set_demography(demography)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{demography}}{(\code{data.frame(1)})\cr The demography of the population per region. The \code{data.frame} must include the following columns:\cr * \code{nuts} (\code{character}) The NUTS 1/2/3 code for the stratification. Only one NUTS level is needed.\cr * \code{...} Optional additional stratification columns (e.g. sex, ethnicity).\cr * \code{population} (\code{numeric}) Number of individuals in the group.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyRegions-describe"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyRegions-describe}{}}}
\subsection{\code{DiseasyRegions$describe()}}{
  Prints a human readable report of the internal state of the module.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyRegions$describe()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyRegions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyRegions-clone}{}}}
\subsection{\code{DiseasyRegions$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DiseasyRegions$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
