% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiseasyVariant.R
\name{DiseasyVariant}
\alias{DiseasyVariant}
\title{TODO}
\description{
TODO
}
\section{Super class}{
\code{\link[diseasy:DiseasyBaseModule]{diseasy::DiseasyBaseModule}} -> \code{DiseasyVariant}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{variants}}{(\code{list})\cr
TODO. Read-only.}

\item{\code{cross_immunity}}{(\code{matrix})\cr
. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DiseasyVariant-new}{\code{DiseasyVariant$new()}}
\item \href{#method-DiseasyVariant-add_variant}{\code{DiseasyVariant$add_variant()}}
\item \href{#method-DiseasyVariant-clone}{\code{DiseasyVariant$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="finalize"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-finalize'><code>diseasy::DiseasyBaseModule$finalize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="load_module"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-load_module'><code>diseasy::DiseasyBaseModule$load_module()</code></a></span></li>
<li><span class="pkg-link" data-pkg="diseasy" data-topic="DiseasyBaseModule" data-id="set_moduleowner"><a href='../../diseasy/html/DiseasyBaseModule.html#method-DiseasyBaseModule-set_moduleowner'><code>diseasy::DiseasyBaseModule$set_moduleowner()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyVariant-new"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyVariant-new}{}}}
\subsection{Method \code{new()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyVariant$new(n_variants = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_variants}}{(\code{numeric(1)})\cr
A quick way to generate a number of variants with decreasing relative infection risk.}

\item{\code{...}}{parameters sent to \code{DiseasyBaseModule} \link[R6:R6Class]{R6} constructor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyVariant-add_variant"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyVariant-add_variant}{}}}
\subsection{Method \code{add_variant()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyVariant$add_variant(name, characteristics)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(\code{character(1)})\cr
The name of the variant.}

\item{\code{characteristics}}{(\code{list})\cr
A named list of characteristics of the variant.
Characteristics can be:
- \code{relative_infection_risk} (\code{numeric(1)}): The relative infection risk of the variant.
- \code{cross_immunity} (\verb{named vector}): The overlap in immunity of when the named variant attempts to infect
a host previously infected by the current variant. If not specified, the default is 1.
- \code{introduction_date} (\code{Date(1)}): The date the variant was introduced into the population.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DiseasyVariant-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DiseasyVariant-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DiseasyVariant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
