% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adjacency_meta.R
\name{generate_adjacency_meta}
\alias{generate_adjacency_meta}
\title{Generate adjacency data with Meta Social Connectedness}
\usage{
generate_adjacency_meta(area = NULL)
}
\arguments{
\item{area}{(\code{character()})\cr The geographic area (list of regions) of interest. If \code{NULL}, all regions are returned.}
}
\value{
(\code{data.frame(1)})\cr The adjacency (connectedness) of the regions. Effectively, the \code{adjacency} is a long-form of the adjacency-matrix\cr The \code{data.frame} must include the following columns:\cr - \code{from} (\code{character}): Region identifier.\cr - \code{to}   (\code{character}): Region identifier.\cr - \code{adjacency} (\code{numeric}): Strength of the connectedness.\cr
}
\description{
Generate adjacency data between NUTS 3 regions from Meta Social Connectedness
Index.
}
\examples{
\dontrun{
adjacency_meta_nordic <- generate_adjacency_meta(area = c("DK", "FI", "IS", "NO", "SE"))
}
}
\keyword{data-generators}
