% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts}
\alias{nuts}
\title{NUTS region identifiers}
\source{
Source: 10.2908/demo_r_pjangrp3.
Retrieved June 15, 2026.
Population on 1 January by age group, sex and NUTS 3 region
}
\description{
This data set contains the unique NUTS region identifiers represented in
\code{demography_nuts} after truncating to the latest available year and removing
rows with missing population values.
}
\details{
This object is derived from \code{demography_nuts}, which is a modified version
of Eurostat dataset \code{demo_r_pjangrp3}. Eurostat is not responsible for the
modifications made to the source data.
}
\author{
Rasmus Skytte Randløv \email{rske@ssi.dk}
}
\keyword{data}
