% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_R6_utils.R
\name{printr}
\alias{printr}
\title{cat printing with default new line}
\usage{
printr(..., file = "/dev/null", sep = "")
}
\arguments{
\item{...}{The normal input to cat}

\item{file}{Path of an output file to append the output to}

\item{sep}{The separator given to cat}
}
\description{
cat printing with default new line
}
